Read capture-recapture data with Headed format used by program E-SURGE

This function reads in capture-recapture dataset with the Headed format. It ignores all forms of censorship for now, and drops continuous covariates because no goodness-of-fit test exists for such models

read_headed(file)

Arguments

file
text file with Headed format

Value

list with first component the matrix of encounter histories, second components the vector of number of individuals with corresponding histories and, if relevant, third component vector/matrix with group(s)

Author

Olivier Gimenez <olivier.gimenez@cefe.cnrs.fr>

Examples

# read in Dipper dataset
dipper = system.file("extdata", "ed.txt", package = "R2ucare")
read_headed(dipper)

#> $encounter_histories
#>        [,1] [,2] [,3] [,4] [,5] [,6] [,7]
#>   [1,]    1    1    1    1    1    1    0
#>   [2,]    1    1    1    1    0    0    0
#>   [3,]    1    1    0    0    0    0    0
#>   [4,]    1    1    0    0    0    0    0
#>   [5,]    1    1    0    0    0    0    0
#>   [6,]    1    1    0    0    0    0    0
#>   [7,]    1    0    1    0    0    0    0
#>   [8,]    1    0    0    0    0    0    0
#>   [9,]    1    0    0    0    0    0    0
#>  [10,]    1    0    0    0    0    0    0
#>  [11,]    1    0    0    0    0    0    0
#>  [12,]    1    0    0    0    0    0    0
#>  [13,]    0    1    1    1    1    0    0
#>  [14,]    0    1    1    1    0    0    0
#>  [15,]    0    1    1    0    0    0    0
#>  [16,]    0    1    1    0    0    0    0
#>  [17,]    0    1    1    0    0    0    0
#>  [18,]    0    1    1    0    0    0    0
#>  [19,]    0    1    1    0    0    0    0
#>  [20,]    0    1    1    0    0    0    0
#>  [21,]    0    1    1    0    0    0    0
#>  [22,]    0    1    0    0    0    0    0
#>  [23,]    0    1    0    0    0    0    0
#>  [24,]    0    1    0    0    0    0    0
#>  [25,]    0    1    0    0    0    0    0
#>  [26,]    0    1    0    0    0    0    0
#>  [27,]    0    1    0    0    0    0    0
#>  [28,]    0    1    0    0    0    0    0
#>  [29,]    0    1    0    0    0    0    0
#>  [30,]    0    1    0    0    0    0    0
#>  [31,]    0    1    0    0    0    0    0
#>  [32,]    0    1    0    0    0    0    0
#>  [33,]    0    0    1    1    1    1    0
#>  [34,]    0    0    1    1    1    0    0
#>  [35,]    0    0    1    1    1    0    0
#>  [36,]    0    0    1    1    1    0    0
#>  [37,]    0    0    1    1    1    0    0
#>  [38,]    0    0    1    1    0    0    0
#>  [39,]    0    0    1    1    0    0    0
#>  [40,]    0    0    1    1    0    0    0
#>  [41,]    0    0    1    1    0    0    0
#>  [42,]    0    0    1    1    0    0    0
#>  [43,]    0    0    1    1    0    0    0
#>  [44,]    0    0    1    1    0    0    0
#>  [45,]    0    0    1    1    0    0    0
#>  [46,]    0    0    1    0    1    1    0
#>  [47,]    0    0    1    0    0    0    0
#>  [48,]    0    0    1    0    0    0    0
#>  [49,]    0    0    1    0    0    0    0
#>  [50,]    0    0    1    0    0    0    0
#>  [51,]    0    0    1    0    0    0    0
#>  [52,]    0    0    1    0    0    0    0
#>  [53,]    0    0    1    0    0    0    0
#>  [54,]    0    0    1    0    0    0    0
#>  [55,]    0    0    1    0    0    0    0
#>  [56,]    0    0    1    0    0    0    0
#>  [57,]    0    0    1    0    0    0    0
#>  [58,]    0    0    0    1    1    1    1
#>  [59,]    0    0    0    1    1    1    1
#>  [60,]    0    0    0    1    1    1    1
#>  [61,]    0    0    0    1    1    1    1
#>  [62,]    0    0    0    1    1    1    1
#>  [63,]    0    0    0    1    1    1    1
#>  [64,]    0    0    0    1    1    1    0
#>  [65,]    0    0    0    1    1    1    0
#>  [66,]    0    0    0    1    1    1    0
#>  [67,]    0    0    0    1    1    0    0
#>  [68,]    0    0    0    1    1    0    0
#>  [69,]    0    0    0    1    1    0    0
#>  [70,]    0    0    0    1    1    0    0
#>  [71,]    0    0    0    1    1    0    0
#>  [72,]    0    0    0    1    1    0    0
#>  [73,]    0    0    0    1    0    0    1
#>  [74,]    0    0    0    1    0    0    0
#>  [75,]    0    0    0    1    0    0    0
#>  [76,]    0    0    0    1    0    0    0
#>  [77,]    0    0    0    1    0    0    0
#>  [78,]    0    0    0    1    0    0    0
#>  [79,]    0    0    0    1    0    0    0
#>  [80,]    0    0    0    0    1    1    1
#>  [81,]    0    0    0    0    1    1    1
#>  [82,]    0    0    0    0    1    1    1
#>  [83,]    0    0    0    0    1    1    1
#>  [84,]    0    0    0    0    1    1    1
#>  [85,]    0    0    0    0    1    1    1
#>  [86,]    0    0    0    0    1    1    1
#>  [87,]    0    0    0    0    1    1    1
#>  [88,]    0    0    0    0    1    1    1
#>  [89,]    0    0    0    0    1    1    1
#>  [90,]    0    0    0    0    1    1    0
#>  [91,]    0    0    0    0    1    1    0
#>  [92,]    0    0    0    0    1    1    0
#>  [93,]    0    0    0    0    1    0    0
#>  [94,]    0    0    0    0    1    0    0
#>  [95,]    0    0    0    0    1    0    0
#>  [96,]    0    0    0    0    1    0    0
#>  [97,]    0    0    0    0    1    0    0
#>  [98,]    0    0    0    0    1    0    0
#>  [99,]    0    0    0    0    1    0    0
#> [100,]    0    0    0    0    1    0    0
#> [101,]    0    0    0    0    1    0    0
#> [102,]    0    0    0    0    0    1    1
#> [103,]    0    0    0    0    0    1    1
#> [104,]    0    0    0    0    0    1    1
#> [105,]    0    0    0    0    0    1    1
#> [106,]    0    0    0    0    0    1    1
#> [107,]    0    0    0    0    0    1    1
#> [108,]    0    0    0    0    0    1    1
#> [109,]    0    0    0    0    0    1    1
#> [110,]    0    0    0    0    0    1    1
#> [111,]    0    0    0    0    0    1    1
#> [112,]    0    0    0    0    0    1    1
#> [113,]    0    0    0    0    0    1    1
#> [114,]    0    0    0    0    0    1    0
#> [115,]    0    0    0    0    0    1    0
#> [116,]    0    0    0    0    0    1    0
#> [117,]    0    0    0    0    0    1    0
#> [118,]    0    0    0    0    0    1    0
#> [119,]    0    0    0    0    0    1    0
#> [120,]    0    0    0    0    0    1    0
#> [121,]    0    0    0    0    0    1    0
#> [122,]    0    0    0    0    0    1    0
#> [123,]    0    0    0    0    0    1    0
#> [124,]    0    0    0    0    0    1    0
#> [125,]    0    0    0    0    0    0    1
#> [126,]    0    0    0    0    0    0    1
#> [127,]    0    0    0    0    0    0    1
#> [128,]    0    0    0    0    0    0    1
#> [129,]    0    0    0    0    0    0    1
#> [130,]    0    0    0    0    0    0    1
#> [131,]    0    0    0    0    0    0    1
#> [132,]    0    0    0    0    0    0    1
#> [133,]    0    0    0    0    0    0    1
#> [134,]    0    0    0    0    0    0    1
#> [135,]    0    0    0    0    0    0    1
#> [136,]    0    0    0    0    0    0    1
#> [137,]    0    0    0    0    0    0    1
#> [138,]    0    0    0    0    0    0    1
#> [139,]    0    0    0    0    0    0    1
#> [140,]    0    0    0    0    0    0    1
#> [141,]    0    0    0    0    0    0    1
#> [142,]    1    1    1    1    1    0    0
#> [143,]    1    1    1    1    0    0    0
#> [144,]    1    1    0    1    1    1    0
#> [145,]    1    1    0    0    0    0    0
#> [146,]    1    1    0    0    0    0    0
#> [147,]    1    0    1    0    0    0    0
#> [148,]    1    0    0    0    0    0    0
#> [149,]    1    0    0    0    0    0    0
#> [150,]    1    0    0    0    0    0    0
#> [151,]    1    0    0    0    0    0    0
#> [152,]    0    1    1    1    1    1    1
#> [153,]    0    1    1    1    1    1    1
#> [154,]    0    1    1    1    1    1    0
#> [155,]    0    1    1    1    1    0    0
#> [156,]    0    1    1    1    1    0    0
#> [157,]    0    1    1    1    0    0    0
#> [158,]    0    1    1    0    1    1    0
#> [159,]    0    1    1    0    0    0    0
#> [160,]    0    1    1    0    0    0    0
#> [161,]    0    1    1    0    0    0    0
#> [162,]    0    1    1    0    0    0    0
#> [163,]    0    1    0    0    0    0    0
#> [164,]    0    1    0    0    0    0    0
#> [165,]    0    1    0    0    0    0    0
#> [166,]    0    1    0    0    0    0    0
#> [167,]    0    1    0    0    0    0    0
#> [168,]    0    1    0    0    0    0    0
#> [169,]    0    1    0    0    0    0    0
#> [170,]    0    1    0    0    0    0    0
#> [171,]    0    1    0    0    0    0    0
#> [172,]    0    1    0    0    0    0    0
#> [173,]    0    1    0    0    0    0    0
#> [174,]    0    1    0    0    0    0    0
#> [175,]    0    1    0    0    0    0    0
#> [176,]    0    1    0    0    0    0    0
#> [177,]    0    1    0    0    0    0    0
#> [178,]    0    1    0    0    0    0    0
#> [179,]    0    1    0    0    0    0    0
#> [180,]    0    1    0    0    0    0    0
#> [181,]    0    0    1    1    1    1    1
#> [182,]    0    0    1    1    1    1    1
#> [183,]    0    0    1    1    1    1    0
#> [184,]    0    0    1    1    1    0    0
#> [185,]    0    0    1    1    1    0    0
#> [186,]    0    0    1    1    0    0    0
#> [187,]    0    0    1    1    0    0    0
#> [188,]    0    0    1    1    0    0    0
#> [189,]    0    0    1    1    0    0    0
#> [190,]    0    0    1    0    0    0    0
#> [191,]    0    0    1    0    0    0    0
#> [192,]    0    0    1    0    0    0    0
#> [193,]    0    0    1    0    0    0    0
#> [194,]    0    0    1    0    0    0    0
#> [195,]    0    0    1    0    0    0    0
#> [196,]    0    0    1    0    0    0    0
#> [197,]    0    0    1    0    0    0    0
#> [198,]    0    0    1    0    0    0    0
#> [199,]    0    0    1    0    0    0    0
#> [200,]    0    0    1    0    0    0    0
#> [201,]    0    0    1    0    0    0    0
#> [202,]    0    0    1    0    0    0    0
#> [203,]    0    0    1    0    0    0    0
#> [204,]    0    0    1    0    0    0    0
#> [205,]    0    0    1    0    0    0    0
#> [206,]    0    0    1    0    0    0    0
#> [207,]    0    0    1    0    0    0    0
#> [208,]    0    0    0    1    1    1    1
#> [209,]    0    0    0    1    1    1    1
#> [210,]    0    0    0    1    1    1    0
#> [211,]    0    0    0    1    1    1    0
#> [212,]    0    0    0    1    1    1    0
#> [213,]    0    0    0    1    1    1    0
#> [214,]    0    0    0    1    1    0    0
#> [215,]    0    0    0    1    1    0    0
#> [216,]    0    0    0    1    1    0    0
#> [217,]    0    0    0    1    1    0    0
#> [218,]    0    0    0    1    1    0    0
#> [219,]    0    0    0    1    0    1    1
#> [220,]    0    0    0    1    0    0    1
#> [221,]    0    0    0    1    0    0    0
#> [222,]    0    0    0    1    0    0    0
#> [223,]    0    0    0    1    0    0    0
#> [224,]    0    0    0    1    0    0    0
#> [225,]    0    0    0    1    0    0    0
#> [226,]    0    0    0    1    0    0    0
#> [227,]    0    0    0    1    0    0    0
#> [228,]    0    0    0    1    0    0    0
#> [229,]    0    0    0    1    0    0    0
#> [230,]    0    0    0    1    0    0    0
#> [231,]    0    0    0    0    1    1    1
#> [232,]    0    0    0    0    1    1    1
#> [233,]    0    0    0    0    1    1    1
#> [234,]    0    0    0    0    1    1    1
#> [235,]    0    0    0    0    1    1    1
#> [236,]    0    0    0    0    1    1    1
#> [237,]    0    0    0    0    1    1    0
#> [238,]    0    0    0    0    1    1    0
#> [239,]    0    0    0    0    1    1    0
#> [240,]    0    0    0    0    1    1    0
#> [241,]    0    0    0    0    1    1    0
#> [242,]    0    0    0    0    1    1    0
#> [243,]    0    0    0    0    1    0    0
#> [244,]    0    0    0    0    1    0    0
#> [245,]    0    0    0    0    1    0    0
#> [246,]    0    0    0    0    1    0    0
#> [247,]    0    0    0    0    1    0    0
#> [248,]    0    0    0    0    1    0    0
#> [249,]    0    0    0    0    1    0    0
#> [250,]    0    0    0    0    0    1    1
#> [251,]    0    0    0    0    0    1    1
#> [252,]    0    0    0    0    0    1    1
#> [253,]    0    0    0    0    0    1    1
#> [254,]    0    0    0    0    0    1    1
#> [255,]    0    0    0    0    0    1    1
#> [256,]    0    0    0    0    0    1    1
#> [257,]    0    0    0    0    0    1    1
#> [258,]    0    0    0    0    0    1    1
#> [259,]    0    0    0    0    0    1    1
#> [260,]    0    0    0    0    0    1    1
#> [261,]    0    0    0    0    0    1    0
#> [262,]    0    0    0    0    0    1    0
#> [263,]    0    0    0    0    0    1    0
#> [264,]    0    0    0    0    0    1    0
#> [265,]    0    0    0    0    0    1    0
#> [266,]    0    0    0    0    0    1    0
#> [267,]    0    0    0    0    0    1    0
#> [268,]    0    0    0    0    0    1    0
#> [269,]    0    0    0    0    0    1    0
#> [270,]    0    0    0    0    0    1    0
#> [271,]    0    0    0    0    0    1    0
#> [272,]    0    0    0    0    0    1    0
#> [273,]    0    0    0    0    0    0    1
#> [274,]    0    0    0    0    0    0    1
#> [275,]    0    0    0    0    0    0    1
#> [276,]    0    0    0    0    0    0    1
#> [277,]    0    0    0    0    0    0    1
#> [278,]    0    0    0    0    0    0    1
#> [279,]    0    0    0    0    0    0    1
#> [280,]    0    0    0    0    0    0    1
#> [281,]    0    0    0    0    0    0    1
#> [282,]    0    0    0    0    0    0    1
#> [283,]    0    0    0    0    0    0    1
#> [284,]    0    0    0    0    0    0    1
#> [285,]    0    0    0    0    0    0    1
#> [286,]    0    0    0    0    0    0    1
#> [287,]    0    0    0    0    0    0    1
#> [288,]    0    0    0    0    0    0    1
#> [289,]    0    0    0    0    0    0    1
#> [290,]    0    0    0    0    0    0    1
#> [291,]    0    0    0    0    0    0    1
#> [292,]    0    0    0    0    0    0    1
#> [293,]    0    0    0    0    0    0    1
#> [294,]    0    0    0    0    0    0    1
#> 
#> $sample_size
#>   [1] 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1
#>  [38] 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1
#>  [75] 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1
#> [112] 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1
#> [149] 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1
#> [186] 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1
#> [223] 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1
#> [260] 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1
#> 
#> $groups
#>   [1] "male"   "male"   "male"   "male"   "male"   "male"   "male"   "male"  
#>   [9] "male"   "male"   "male"   "male"   "male"   "male"   "male"   "male"  
#>  [17] "male"   "male"   "male"   "male"   "male"   "male"   "male"   "male"  
#>  [25] "male"   "male"   "male"   "male"   "male"   "male"   "male"   "male"  
#>  [33] "male"   "male"   "male"   "male"   "male"   "male"   "male"   "male"  
#>  [41] "male"   "male"   "male"   "male"   "male"   "male"   "male"   "male"  
#>  [49] "male"   "male"   "male"   "male"   "male"   "male"   "male"   "male"  
#>  [57] "male"   "male"   "male"   "male"   "male"   "male"   "male"   "male"  
#>  [65] "male"   "male"   "male"   "male"   "male"   "male"   "male"   "male"  
#>  [73] "male"   "male"   "male"   "male"   "male"   "male"   "male"   "male"  
#>  [81] "male"   "male"   "male"   "male"   "male"   "male"   "male"   "male"  
#>  [89] "male"   "male"   "male"   "male"   "male"   "male"   "male"   "male"  
#>  [97] "male"   "male"   "male"   "male"   "male"   "male"   "male"   "male"  
#> [105] "male"   "male"   "male"   "male"   "male"   "male"   "male"   "male"  
#> [113] "male"   "male"   "male"   "male"   "male"   "male"   "male"   "male"  
#> [121] "male"   "male"   "male"   "male"   "male"   "male"   "male"   "male"  
#> [129] "male"   "male"   "male"   "male"   "male"   "male"   "male"   "male"  
#> [137] "male"   "male"   "male"   "male"   "male"   "female" "female" "female"
#> [145] "female" "female" "female" "female" "female" "female" "female" "female"
#> [153] "female" "female" "female" "female" "female" "female" "female" "female"
#> [161] "female" "female" "female" "female" "female" "female" "female" "female"
#> [169] "female" "female" "female" "female" "female" "female" "female" "female"
#> [177] "female" "female" "female" "female" "female" "female" "female" "female"
#> [185] "female" "female" "female" "female" "female" "female" "female" "female"
#> [193] "female" "female" "female" "female" "female" "female" "female" "female"
#> [201] "female" "female" "female" "female" "female" "female" "female" "female"
#> [209] "female" "female" "female" "female" "female" "female" "female" "female"
#> [217] "female" "female" "female" "female" "female" "female" "female" "female"
#> [225] "female" "female" "female" "female" "female" "female" "female" "female"
#> [233] "female" "female" "female" "female" "female" "female" "female" "female"
#> [241] "female" "female" "female" "female" "female" "female" "female" "female"
#> [249] "female" "female" "female" "female" "female" "female" "female" "female"
#> [257] "female" "female" "female" "female" "female" "female" "female" "female"
#> [265] "female" "female" "female" "female" "female" "female" "female" "female"
#> [273] "female" "female" "female" "female" "female" "female" "female" "female"
#> [281] "female" "female" "female" "female" "female" "female" "female" "female"
#> [289] "female" "female" "female" "female" "female" "female"
#> 

# read in Geese dataset
geese = system.file("extdata", "geese.txt", package = "R2ucare")
read_headed(geese)

#> $encounter_histories
#>        H.A1984 H. H..1 H..2 H..3 H.A1989
#>   [1,]       1  1    1    1    1       2
#>   [2,]       1  1    1    1    1       3
#>   [3,]       1  1    1    1    1       4
#>   [4,]       1  1    1    1    2       1
#>   [5,]       1  1    1    1    2       2
#>   [6,]       1  1    1    1    2       3
#>   [7,]       1  1    1    1    3       1
#>   [8,]       1  1    1    1    3       2
#>   [9,]       1  1    1    1    3       3
#>  [10,]       1  1    1    1    3       4
#>  [11,]       1  1    1    1    4       1
#>  [12,]       1  1    1    1    4       2
#>  [13,]       1  1    1    1    4       3
#>  [14,]       1  1    1    1    4       4
#>  [15,]       1  1    1    2    1       1
#>  [16,]       1  1    1    2    1       2
#>  [17,]       1  1    1    2    1       3
#>  [18,]       1  1    1    2    2       1
#>  [19,]       1  1    1    2    2       2
#>  [20,]       1  1    1    2    2       3
#>  [21,]       1  1    1    2    3       1
#>  [22,]       1  1    1    2    3       2
#>  [23,]       1  1    1    2    3       3
#>  [24,]       1  1    1    2    3       4
#>  [25,]       1  1    1    2    4       1
#>  [26,]       1  1    1    2    4       3
#>  [27,]       1  1    1    3    1       1
#>  [28,]       1  1    1    3    1       2
#>  [29,]       1  1    1    3    1       3
#>  [30,]       1  1    1    3    1       4
#>  [31,]       1  1    1    3    2       1
#>  [32,]       1  1    1    3    2       2
#>  [33,]       1  1    1    3    2       3
#>  [34,]       1  1    1    3    3       1
#>  [35,]       1  1    1    3    3       2
#>  [36,]       1  1    1    3    3       3
#>  [37,]       1  1    1    3    3       4
#>  [38,]       1  1    1    3    4       1
#>  [39,]       1  1    1    3    4       2
#>  [40,]       1  1    1    3    4       3
#>  [41,]       1  1    1    3    4       4
#>  [42,]       1  1    1    4    1       1
#>  [43,]       1  1    1    4    1       2
#>  [44,]       1  1    1    4    1       3
#>  [45,]       1  1    1    4    1       4
#>  [46,]       1  1    1    4    2       1
#>  [47,]       1  1    1    4    2       2
#>  [48,]       1  1    1    4    2       4
#>  [49,]       1  1    1    4    3       1
#>  [50,]       1  1    1    4    3       2
#>  [51,]       1  1    1    4    3       3
#>  [52,]       1  1    1    4    3       4
#>  [53,]       1  1    1    4    4       1
#>  [54,]       1  1    1    4    4       2
#>  [55,]       1  1    1    4    4       3
#>  [56,]       1  1    1    4    4       4
#>  [57,]       1  1    2    1    1       1
#>  [58,]       1  1    2    1    1       2
#>  [59,]       1  1    2    1    1       3
#>  [60,]       1  1    2    1    1       4
#>  [61,]       1  1    2    1    2       1
#>  [62,]       1  1    2    1    2       2
#>  [63,]       1  1    2    1    2       3
#>  [64,]       1  1    2    1    3       1
#>  [65,]       1  1    2    1    3       2
#>  [66,]       1  1    2    1    3       3
#>  [67,]       1  1    2    1    4       2
#>  [68,]       1  1    2    1    4       3
#>  [69,]       1  1    2    1    4       4
#>  [70,]       1  1    2    2    1       1
#>  [71,]       1  1    2    2    1       2
#>  [72,]       1  1    2    2    1       3
#>  [73,]       1  1    2    2    2       1
#>  [74,]       1  1    2    2    2       2
#>  [75,]       1  1    2    2    2       3
#>  [76,]       1  1    2    2    3       1
#>  [77,]       1  1    2    2    3       2
#>  [78,]       1  1    2    2    3       3
#>  [79,]       1  1    2    3    1       1
#>  [80,]       1  1    2    3    1       2
#>  [81,]       1  1    2    3    1       3
#>  [82,]       1  1    2    3    2       1
#>  [83,]       1  1    2    3    2       2
#>  [84,]       1  1    2    3    2       3
#>  [85,]       1  1    2    3    3       1
#>  [86,]       1  1    2    3    3       2
#>  [87,]       1  1    2    3    3       3
#>  [88,]       1  1    2    3    4       1
#>  [89,]       1  1    2    4    1       1
#>  [90,]       1  1    2    4    2       1
#>  [91,]       1  1    3    1    1       1
#>  [92,]       1  1    3    1    1       2
#>  [93,]       1  1    3    1    1       3
#>  [94,]       1  1    3    1    1       4
#>  [95,]       1  1    3    1    2       1
#>  [96,]       1  1    3    1    2       2
#>  [97,]       1  1    3    1    2       3
#>  [98,]       1  1    3    1    3       1
#>  [99,]       1  1    3    1    3       2
#> [100,]       1  1    3    1    3       3
#> [101,]       1  1    3    1    3       4
#> [102,]       1  1    3    1    4       1
#> [103,]       1  1    3    1    4       3
#> [104,]       1  1    3    2    1       1
#> [105,]       1  1    3    2    1       2
#> [106,]       1  1    3    2    1       3
#> [107,]       1  1    3    2    2       1
#> [108,]       1  1    3    2    2       2
#> [109,]       1  1    3    2    2       3
#> [110,]       1  1    3    2    3       1
#> [111,]       1  1    3    2    3       2
#> [112,]       1  1    3    2    3       3
#> [113,]       1  1    3    2    4       2
#> [114,]       1  1    3    3    1       1
#> [115,]       1  1    3    3    1       2
#> [116,]       1  1    3    3    1       3
#> [117,]       1  1    3    3    2       1
#> [118,]       1  1    3    3    2       2
#> [119,]       1  1    3    3    2       3
#> [120,]       1  1    3    3    3       1
#> [121,]       1  1    3    3    3       2
#> [122,]       1  1    3    3    3       3
#> [123,]       1  1    3    3    4       1
#> [124,]       1  1    3    3    4       3
#> [125,]       1  1    3    4    1       1
#> [126,]       1  1    3    4    3       1
#> [127,]       1  1    3    4    3       2
#> [128,]       1  1    3    4    4       1
#> [129,]       1  1    4    1    1       1
#> [130,]       1  1    4    1    1       2
#> [131,]       1  1    4    1    1       3
#> [132,]       1  1    4    1    1       4
#> [133,]       1  1    4    1    2       1
#> [134,]       1  1    4    1    2       3
#> [135,]       1  1    4    1    3       1
#> [136,]       1  1    4    1    3       3
#> [137,]       1  1    4    1    3       4
#> [138,]       1  1    4    1    4       1
#> [139,]       1  1    4    1    4       3
#> [140,]       1  1    4    1    4       4
#> [141,]       1  1    4    2    1       1
#> [142,]       1  1    4    3    1       1
#> [143,]       1  1    4    3    1       3
#> [144,]       1  1    4    3    3       1
#> [145,]       1  1    4    3    3       2
#> [146,]       1  1    4    3    3       3
#> [147,]       1  1    4    3    4       1
#> [148,]       1  1    4    3    4       4
#> [149,]       1  1    4    4    1       1
#> [150,]       1  1    4    4    1       2
#> [151,]       1  1    4    4    1       3
#> [152,]       1  1    4    4    1       4
#> [153,]       1  1    4    4    3       1
#> [154,]       1  1    4    4    4       1
#> [155,]       1  1    4    4    4       3
#> [156,]       1  1    4    4    4       4
#> [157,]       1  2    1    1    1       1
#> [158,]       1  2    1    1    1       2
#> [159,]       1  2    1    1    1       3
#> [160,]       1  2    1    1    2       1
#> [161,]       1  2    1    1    2       2
#> [162,]       1  2    1    1    2       3
#> [163,]       1  2    1    1    3       1
#> [164,]       1  2    1    1    3       2
#> [165,]       1  2    1    1    3       3
#> [166,]       1  2    1    1    4       1
#> [167,]       1  2    1    2    1       1
#> [168,]       1  2    1    2    1       2
#> [169,]       1  2    1    2    1       3
#> [170,]       1  2    1    2    2       1
#> [171,]       1  2    1    2    2       2
#> [172,]       1  2    1    2    3       1
#> [173,]       1  2    1    3    1       1
#> [174,]       1  2    1    3    1       2
#> [175,]       1  2    1    3    1       3
#> [176,]       1  2    1    3    2       1
#> [177,]       1  2    1    3    2       3
#> [178,]       1  2    1    3    3       1
#> [179,]       1  2    1    3    3       2
#> [180,]       1  2    1    3    3       3
#> [181,]       1  2    1    4    1       1
#> [182,]       1  2    1    4    4       1
#> [183,]       1  2    2    1    1       1
#> [184,]       1  2    2    1    1       2
#> [185,]       1  2    2    1    1       3
#> [186,]       1  2    2    1    2       1
#> [187,]       1  2    2    1    2       2
#> [188,]       1  2    2    1    2       3
#> [189,]       1  2    2    1    3       1
#> [190,]       1  2    2    1    3       3
#> [191,]       1  2    2    2    1       1
#> [192,]       1  2    2    2    1       2
#> [193,]       1  2    2    2    1       3
#> [194,]       1  2    2    2    2       1
#> [195,]       1  2    2    2    2       2
#> [196,]       1  2    2    2    2       3
#> [197,]       1  2    2    2    3       1
#> [198,]       1  2    2    3    1       1
#> [199,]       1  2    2    3    2       1
#> [200,]       1  2    2    3    3       1
#> [201,]       1  2    2    3    3       2
#> [202,]       1  2    3    1    1       1
#> [203,]       1  2    3    1    1       2
#> [204,]       1  2    3    1    1       3
#> [205,]       1  2    3    1    2       1
#> [206,]       1  2    3    1    2       2
#> [207,]       1  2    3    1    3       1
#> [208,]       1  2    3    1    3       3
#> [209,]       1  2    3    2    1       1
#> [210,]       1  2    3    2    2       1
#> [211,]       1  2    3    2    2       2
#> [212,]       1  2    3    3    1       1
#> [213,]       1  2    3    3    1       3
#> [214,]       1  2    3    3    3       1
#> [215,]       1  2    3    3    3       2
#> [216,]       1  2    3    4    1       1
#> [217,]       1  3    1    1    1       1
#> [218,]       1  3    1    1    1       2
#> [219,]       1  3    1    1    1       3
#> [220,]       1  3    1    1    1       4
#> [221,]       1  3    1    1    2       1
#> [222,]       1  3    1    1    2       2
#> [223,]       1  3    1    1    2       3
#> [224,]       1  3    1    1    3       1
#> [225,]       1  3    1    1    3       2
#> [226,]       1  3    1    1    3       3
#> [227,]       1  3    1    1    4       1
#> [228,]       1  3    1    1    4       3
#> [229,]       1  3    1    2    1       1
#> [230,]       1  3    1    2    1       2
#> [231,]       1  3    1    2    1       3
#> [232,]       1  3    1    2    2       1
#> [233,]       1  3    1    2    2       2
#> [234,]       1  3    1    2    2       3
#> [235,]       1  3    1    2    3       1
#> [236,]       1  3    1    2    3       2
#> [237,]       1  3    1    2    3       3
#> [238,]       1  3    1    3    1       1
#> [239,]       1  3    1    3    1       2
#> [240,]       1  3    1    3    1       3
#> [241,]       1  3    1    3    1       4
#> [242,]       1  3    1    3    2       1
#> [243,]       1  3    1    3    2       3
#> [244,]       1  3    1    3    3       1
#> [245,]       1  3    1    3    3       3
#> [246,]       1  3    1    3    4       1
#> [247,]       1  3    1    4    1       1
#> [248,]       1  3    1    4    1       4
#> [249,]       1  3    1    4    2       1
#> [250,]       1  3    2    1    1       1
#> [251,]       1  3    2    1    1       3
#> [252,]       1  3    2    1    2       1
#> [253,]       1  3    2    1    2       2
#> [254,]       1  3    2    1    2       3
#> [255,]       1  3    2    1    3       1
#> [256,]       1  3    2    1    3       2
#> [257,]       1  3    2    1    3       3
#> [258,]       1  3    2    2    1       1
#> [259,]       1  3    2    2    2       1
#> [260,]       1  3    2    2    2       2
#> [261,]       1  3    2    2    2       3
#> [262,]       1  3    2    2    3       1
#> [263,]       1  3    2    2    3       2
#> [264,]       1  3    2    3    1       1
#> [265,]       1  3    2    3    1       3
#> [266,]       1  3    2    3    2       1
#> [267,]       1  3    2    3    3       1
#> [268,]       1  3    2    3    3       3
#> [269,]       1  3    2    4    1       3
#> [270,]       1  3    2    4    2       1
#> [271,]       1  3    3    1    1       1
#> [272,]       1  3    3    1    1       2
#> [273,]       1  3    3    1    1       3
#> [274,]       1  3    3    1    2       1
#> [275,]       1  3    3    1    2       2
#> [276,]       1  3    3    1    2       3
#> [277,]       1  3    3    1    3       1
#> [278,]       1  3    3    1    3       2
#> [279,]       1  3    3    1    3       3
#> [280,]       1  3    3    1    4       1
#> [281,]       1  3    3    1    4       4
#> [282,]       1  3    3    2    1       1
#> [283,]       1  3    3    2    1       3
#> [284,]       1  3    3    2    2       1
#> [285,]       1  3    3    2    3       1
#> [286,]       1  3    3    3    1       1
#> [287,]       1  3    3    3    1       2
#> [288,]       1  3    3    3    1       3
#> [289,]       1  3    3    3    2       1
#> [290,]       1  3    3    3    2       2
#> [291,]       1  3    3    3    3       1
#> [292,]       1  3    3    3    3       3
#> [293,]       1  3    3    4    1       1
#> [294,]       1  3    3    4    1       4
#> [295,]       1  3    4    1    1       1
#> [296,]       1  3    4    1    1       3
#> [297,]       1  3    4    1    1       4
#> [298,]       1  3    4    1    3       1
#> [299,]       1  3    4    4    1       1
#> [300,]       1  3    4    4    4       1
#> [301,]       1  4    1    1    1       1
#> [302,]       1  4    1    1    1       2
#> [303,]       1  4    1    1    1       3
#> [304,]       1  4    1    1    1       4
#> [305,]       1  4    1    1    2       1
#> [306,]       1  4    1    1    2       2
#> [307,]       1  4    1    1    2       3
#> [308,]       1  4    1    1    3       1
#> [309,]       1  4    1    1    3       3
#> [310,]       1  4    1    1    3       4
#> [311,]       1  4    1    1    4       1
#> [312,]       1  4    1    1    4       3
#> [313,]       1  4    1    1    4       4
#> [314,]       1  4    1    2    1       1
#> [315,]       1  4    1    2    1       2
#> [316,]       1  4    1    2    3       1
#> [317,]       1  4    1    3    1       1
#> [318,]       1  4    1    3    1       3
#> [319,]       1  4    1    3    1       4
#> [320,]       1  4    1    3    3       1
#> [321,]       1  4    1    3    3       3
#> [322,]       1  4    1    3    4       1
#> [323,]       1  4    1    4    1       1
#> [324,]       1  4    1    4    1       2
#> [325,]       1  4    1    4    1       4
#> [326,]       1  4    1    4    3       4
#> [327,]       1  4    1    4    4       1
#> [328,]       1  4    1    4    4       4
#> [329,]       1  4    2    1    1       1
#> [330,]       1  4    2    1    1       3
#> [331,]       1  4    2    2    2       1
#> [332,]       1  4    2    3    1       1
#> [333,]       1  4    2    3    2       1
#> [334,]       1  4    2    4    1       2
#> [335,]       1  4    3    1    1       1
#> [336,]       1  4    3    1    1       3
#> [337,]       1  4    3    1    1       4
#> [338,]       1  4    3    1    2       1
#> [339,]       1  4    3    1    3       1
#> [340,]       1  4    3    2    2       1
#> [341,]       1  4    3    2    3       1
#> [342,]       1  4    3    3    1       1
#> [343,]       1  4    3    3    1       3
#> [344,]       1  4    3    3    2       1
#> [345,]       1  4    3    3    3       1
#> [346,]       1  4    3    3    3       3
#> [347,]       1  4    3    4    1       1
#> [348,]       1  4    3    4    4       1
#> [349,]       1  4    3    4    4       3
#> [350,]       1  4    4    1    1       1
#> [351,]       1  4    4    1    1       3
#> [352,]       1  4    4    1    1       4
#> [353,]       1  4    4    1    2       1
#> [354,]       1  4    4    1    3       1
#> [355,]       1  4    4    1    4       4
#> [356,]       1  4    4    2    1       3
#> [357,]       1  4    4    3    1       1
#> [358,]       1  4    4    3    1       3
#> [359,]       1  4    4    3    3       1
#> [360,]       1  4    4    3    4       1
#> [361,]       1  4    4    3    4       4
#> [362,]       1  4    4    4    1       1
#> [363,]       1  4    4    4    1       4
#> [364,]       1  4    4    4    3       1
#> [365,]       1  4    4    4    4       1
#> [366,]       1  4    4    4    4       4
#> [367,]       2  1    1    1    1       1
#> [368,]       2  1    1    1    1       2
#> [369,]       2  1    1    1    1       3
#> [370,]       2  1    1    1    2       1
#> [371,]       2  1    1    1    3       1
#> [372,]       2  1    1    1    3       3
#> [373,]       2  1    1    1    4       1
#> [374,]       2  1    1    2    1       1
#> [375,]       2  1    1    2    1       2
#> [376,]       2  1    1    2    2       1
#> [377,]       2  1    1    2    2       2
#> [378,]       2  1    1    2    3       1
#> [379,]       2  1    1    3    1       1
#> [380,]       2  1    1    3    1       3
#> [381,]       2  1    2    1    1       1
#> [382,]       2  1    2    1    1       3
#> [383,]       2  1    2    1    2       1
#> [384,]       2  1    2    1    3       1
#> [385,]       2  1    2    2    1       1
#> [386,]       2  1    2    2    1       2
#> [387,]       2  1    2    2    1       3
#> [388,]       2  1    2    3    2       2
#> [389,]       2  1    3    1    1       1
#> [390,]       2  1    3    1    1       2
#> [391,]       2  1    3    1    2       1
#> [392,]       2  1    3    1    3       3
#> [393,]       2  1    3    2    1       2
#> [394,]       2  1    3    2    3       1
#> [395,]       2  1    3    3    1       1
#> [396,]       2  1    3    3    3       1
#> [397,]       2  2    1    1    1       1
#> [398,]       2  2    1    1    1       2
#> [399,]       2  2    1    1    2       1
#> [400,]       2  2    1    1    3       1
#> [401,]       2  2    1    2    1       1
#> [402,]       2  2    1    2    1       2
#> [403,]       2  2    1    2    2       1
#> [404,]       2  2    1    2    2       2
#> [405,]       2  2    1    3    1       1
#> [406,]       2  2    1    3    1       2
#> [407,]       2  2    2    1    1       1
#> [408,]       2  2    2    1    1       2
#> [409,]       2  2    2    1    2       1
#> [410,]       2  2    2    1    2       2
#> [411,]       2  2    2    1    2       3
#> [412,]       2  2    2    1    3       1
#> [413,]       2  2    2    2    1       1
#> [414,]       2  2    2    2    1       2
#> [415,]       2  2    2    2    2       1
#> [416,]       2  2    2    2    2       2
#> [417,]       2  2    2    2    3       1
#> [418,]       2  2    2    3    1       1
#> [419,]       2  2    2    3    1       2
#> [420,]       2  2    3    1    1       1
#> [421,]       2  3    1    1    1       1
#> [422,]       2  3    1    1    1       3
#> [423,]       2  3    1    1    3       1
#> [424,]       2  3    1    2    1       1
#> [425,]       2  3    1    3    1       2
#> [426,]       2  3    2    1    1       1
#> [427,]       2  3    2    1    1       2
#> [428,]       2  3    2    2    1       1
#> [429,]       2  3    2    2    2       1
#> [430,]       2  3    2    2    2       2
#> [431,]       2  3    2    3    3       1
#> [432,]       2  3    3    1    1       1
#> [433,]       3  1    1    1    1       1
#> [434,]       3  1    1    1    1       2
#> [435,]       3  1    1    1    1       3
#> [436,]       3  1    1    1    2       1
#> [437,]       3  1    1    1    2       2
#> [438,]       3  1    1    1    2       3
#> [439,]       3  1    1    1    3       1
#> [440,]       3  1    1    1    3       2
#> [441,]       3  1    1    1    3       3
#> [442,]       3  1    1    1    3       4
#> [443,]       3  1    1    1    4       1
#> [444,]       3  1    1    2    1       1
#> [445,]       3  1    1    2    1       3
#> [446,]       3  1    1    2    3       1
#> [447,]       3  1    1    2    4       3
#> [448,]       3  1    1    3    1       1
#> [449,]       3  1    1    3    1       2
#> [450,]       3  1    1    3    1       3
#> [451,]       3  1    1    3    3       1
#> [452,]       3  1    1    3    3       3
#> [453,]       3  1    1    4    1       1
#> [454,]       3  1    1    4    3       2
#> [455,]       3  1    1    4    4       4
#> [456,]       3  1    2    1    1       1
#> [457,]       3  1    2    1    1       2
#> [458,]       3  1    2    1    2       2
#> [459,]       3  1    2    2    1       1
#> [460,]       3  1    2    2    1       2
#> [461,]       3  1    2    2    1       3
#> [462,]       3  1    2    2    2       1
#> [463,]       3  1    2    3    1       1
#> [464,]       3  1    2    3    1       2
#> [465,]       3  1    2    3    1       3
#> [466,]       3  1    2    3    3       1
#> [467,]       3  1    3    1    1       1
#> [468,]       3  1    3    1    1       2
#> [469,]       3  1    3    1    1       3
#> [470,]       3  1    3    1    2       2
#> [471,]       3  1    3    1    2       3
#> [472,]       3  1    3    1    3       1
#> [473,]       3  1    3    1    3       3
#> [474,]       3  1    3    2    1       1
#> [475,]       3  1    3    2    2       1
#> [476,]       3  1    3    2    3       3
#> [477,]       3  1    3    3    1       1
#> [478,]       3  1    3    3    1       2
#> [479,]       3  1    3    3    1       3
#> [480,]       3  1    3    3    3       1
#> [481,]       3  1    3    3    3       3
#> [482,]       3  1    3    3    4       1
#> [483,]       3  1    3    4    1       1
#> [484,]       3  1    4    1    1       4
#> [485,]       3  1    4    4    1       1
#> [486,]       3  2    1    1    1       1
#> [487,]       3  2    1    1    1       2
#> [488,]       3  2    1    1    1       3
#> [489,]       3  2    1    1    3       1
#> [490,]       3  2    1    1    3       3
#> [491,]       3  2    1    2    1       1
#> [492,]       3  2    1    2    1       2
#> [493,]       3  2    1    2    2       1
#> [494,]       3  2    1    2    3       2
#> [495,]       3  2    1    3    1       1
#> [496,]       3  2    1    3    1       3
#> [497,]       3  2    1    3    2       1
#> [498,]       3  2    2    1    1       1
#> [499,]       3  2    2    2    1       1
#> [500,]       3  2    2    2    1       2
#> [501,]       3  2    2    2    1       3
#> [502,]       3  2    2    2    2       1
#> [503,]       3  2    2    3    1       1
#> [504,]       3  2    2    4    1       1
#> [505,]       3  2    3    1    1       1
#> [506,]       3  3    1    1    1       1
#> [507,]       3  3    1    1    1       2
#> [508,]       3  3    1    1    1       3
#> [509,]       3  3    1    1    1       4
#> [510,]       3  3    1    1    2       1
#> [511,]       3  3    1    1    2       2
#> [512,]       3  3    1    1    2       3
#> [513,]       3  3    1    1    3       1
#> [514,]       3  3    1    1    3       2
#> [515,]       3  3    1    1    3       3
#> [516,]       3  3    1    1    4       1
#> [517,]       3  3    1    2    1       1
#> [518,]       3  3    1    2    2       1
#> [519,]       3  3    1    2    2       2
#> [520,]       3  3    1    3    1       1
#> [521,]       3  3    1    3    1       2
#> [522,]       3  3    1    3    1       3
#> [523,]       3  3    1    3    2       1
#> [524,]       3  3    1    3    3       1
#> [525,]       3  3    1    3    3       2
#> [526,]       3  3    1    3    3       3
#> [527,]       3  3    1    3    4       3
#> [528,]       3  3    2    1    1       1
#> [529,]       3  3    2    1    1       2
#> [530,]       3  3    2    1    1       3
#> [531,]       3  3    2    2    1       1
#> [532,]       3  3    2    3    1       1
#> [533,]       3  3    3    1    1       1
#> [534,]       3  3    3    1    1       2
#> [535,]       3  3    3    1    1       3
#> [536,]       3  3    3    1    3       1
#> [537,]       3  3    3    1    4       1
#> [538,]       3  3    3    3    1       1
#> [539,]       3  3    3    3    1       2
#> [540,]       3  3    3    3    1       3
#> [541,]       3  3    3    3    2       1
#> [542,]       3  3    3    3    3       1
#> [543,]       3  3    3    3    3       3
#> [544,]       3  3    4    1    1       1
#> [545,]       3  4    1    1    1       1
#> [546,]       3  4    1    1    3       1
#> [547,]       3  4    2    1    1       1
#> [548,]       3  4    3    1    1       2
#> [549,]       4  1    1    1    1       1
#> [550,]       4  1    1    1    1       2
#> [551,]       4  1    1    1    1       3
#> [552,]       4  1    1    1    1       4
#> [553,]       4  1    1    1    3       1
#> [554,]       4  1    1    1    3       3
#> [555,]       4  1    1    1    4       1
#> [556,]       4  1    1    2    1       1
#> [557,]       4  1    1    3    1       1
#> [558,]       4  1    1    3    1       3
#> [559,]       4  1    1    3    3       1
#> [560,]       4  1    1    3    3       3
#> [561,]       4  1    1    4    1       1
#> [562,]       4  1    1    4    4       1
#> [563,]       4  1    2    1    1       1
#> [564,]       4  1    2    1    3       1
#> [565,]       4  1    2    1    4       1
#> [566,]       4  1    2    2    1       1
#> [567,]       4  1    3    1    1       1
#> [568,]       4  1    3    1    1       3
#> [569,]       4  1    3    3    1       1
#> [570,]       4  1    3    3    1       3
#> [571,]       4  1    3    3    3       1
#> [572,]       4  1    4    1    1       1
#> [573,]       4  1    4    1    1       4
#> [574,]       4  1    4    1    4       1
#> [575,]       4  1    4    1    4       4
#> [576,]       4  1    4    4    1       1
#> [577,]       4  1    4    4    1       3
#> [578,]       4  1    4    4    1       4
#> [579,]       4  1    4    4    4       1
#> [580,]       4  1    4    4    4       4
#> [581,]       4  2    1    1    1       1
#> [582,]       4  2    1    1    1       2
#> [583,]       4  2    1    1    1       3
#> [584,]       4  2    1    1    3       1
#> [585,]       4  2    1    1    3       2
#> [586,]       4  2    1    2    1       1
#> [587,]       4  2    1    2    2       2
#> [588,]       4  2    2    1    1       1
#> [589,]       4  2    2    1    2       1
#> [590,]       4  2    2    1    3       2
#> [591,]       4  2    2    2    1       1
#> [592,]       4  2    3    2    2       2
#> [593,]       4  3    1    1    1       1
#> [594,]       4  3    1    1    1       3
#> [595,]       4  3    1    1    3       1
#> [596,]       4  3    1    3    1       1
#> [597,]       4  3    1    3    2       3
#> [598,]       4  3    1    4    3       1
#> [599,]       4  3    2    1    1       1
#> [600,]       4  3    3    1    1       1
#> [601,]       4  3    3    1    1       3
#> [602,]       4  3    3    3    1       1
#> [603,]       4  3    3    3    1       2
#> [604,]       4  3    4    1    1       1
#> [605,]       4  4    1    1    1       1
#> [606,]       4  4    1    1    1       4
#> [607,]       4  4    1    1    4       1
#> [608,]       4  4    1    3    1       1
#> [609,]       4  4    1    3    4       1
#> [610,]       4  4    1    4    1       1
#> [611,]       4  4    2    1    1       1
#> [612,]       4  4    3    1    1       1
#> [613,]       4  4    3    1    1       3
#> [614,]       4  4    3    1    3       4
#> [615,]       4  4    3    3    1       1
#> [616,]       4  4    3    4    4       1
#> [617,]       4  4    4    1    1       1
#> [618,]       4  4    4    1    4       4
#> [619,]       4  4    4    3    1       3
#> [620,]       4  4    4    4    1       1
#> [621,]       4  4    4    4    1       2
#> [622,]       4  4    4    4    1       4
#> [623,]       4  4    4    4    4       4
#> 
#> $sample_size
#>   [1]  158  352  271  317   62   31  748   56  234    7  504   14   71   72  643
#>  [16]   57   36   98   42   10   70   13   25    1    1    2 1040   27  117   11
#>  [31]   59    2   12  285   11  100    2   10    1    5    2  694    7   66   60
#>  [46]   13    1    2   59    2   13    2   86    1   12   31 1135   40   39    2
#>  [61]   63   22    8   68    8   13    1    2    1  201   25   11   56   38    1
#>  [76]   21    4    1   81    9   12   11    5    4   22    6    4    1    3    1
#>  [91] 1559   18   85    4   33    6    3  163    3   51    1    7    1   55    6
#> [106]    2    9    1    4    7    1    4    1  423   10   45    5    5    4  121
#> [121]    2   48    6    1   16    4    1    1  473    1   17   21    9    1   19
#> [136]    2    1   20    1    7    6   22    3    5    1    1    2    1   57    2
#> [151]    2   15    3   10    3   13  920   18    9   36    9    1   24    5    6
#> [166]    3   83    7    1   22    7    2   44    2    7    1    1    5    2    2
#> [181]    2    1  195    8    2   29    8    1    5    1   52    6    1   41   24
#> [196]    2    5    6    2    1    1   73    3    3    2    1    4    2    8    1
#> [211]    1   19    3    2    1    1 1769   17   56    1   28    2    4  103    7
#> [226]   29    1    1   39    2    1    4    1    1    3    1    1  168    3   20
#> [241]    1    4    1   48   15    3    7    1    1   65    2    2    2    1   14
#> [256]    2    1   12    3    3    1    4    1    8    1    2    4    2    1    1
#> [271]  391    3   35    1    1    2   49    2   22    4    1   10    1    2    2
#> [286]   89    3   15    1    3   38   36    1    1    6    2    1    1    2    1
#> [301]  623    1   14    9    5    1    1   12    2    1    9    3    5    6    1
#> [316]    1   27    2    2    7    1    5   26    1    8    1    2    8    7    1
#> [331]    1    1    1    1   51    3    2    1    4    1    1   12    2    2    3
#> [346]    2    7    1    1   56    2    2    1    4    8    1    5    1    1    1
#> [361]    1   25   11    2    4   11  401    5    2    6    4    1    1    6    3
#> [376]    1    2    1    4    2   23    1    1    1    5    3    1    1    9    1
#> [391]    1    1    1    1    3    1  128    1    5    2   12    3    4    2    2
#> [406]    2   40    1    5    4    1    1   10    1    5    4    1    1    1    3
#> [421]   32    1    1    2    1    5    1    1    2    1    1    5 1001    3   22
#> [436]    3    3    1   26    1    4    1    2   18    1    2    1   60    1   11
#> [451]   16    4    1    1    1   19    1    2    2    1    1    1    6    1    1
#> [466]    1   82    2   12    1    1   11    4    3    1    1   23    1    7    6
#> [481]    1    1    1    1    1   47    1    3    1    1    1    1    1    1    1
#> [496]    1    1   10    1    4    1    1    1    1    6  328    1   16    1    5
#> [511]    1    1   18    1    3    1    5    2    1   41    1    6    3   11    1
#> [526]    5    1    9    1    1    2    2   73    1    4    7    1   39    1    5
#> [541]    1    6    8    1    3    1    1    1  380    2    5    3    6    2    3
#> [556]    3    7    1    1    1    7    3    8    1    1    1   19    2    7    1
#> [571]    1    9    1    1    1    2    1    1    1    1    9    1    1    2    1
#> [586]    1    1    4    1    1    1    1   36    1    2    1    1    1    1    2
#> [601]    1    1    1    1   29    2    1    2    1    2    1    2    1    1    2
#> [616]    1   10    1    1    4    1    2    3
#>