Read capture-recapture data with Input (.inp) format used by program MARK

This function reads in capture-recapture dataset with the Input format. It is a wrapper for the function convert.inp from package RMark. It drops continuous covariates because no goodness-of-fit test exists for such models

read_inp(file, group.df = NULL)

Arguments

file
text file with Input format (extension .inp)

group.df
dataframe with grouping variables; contains a row for each group defined in the input file row1=group1, row2=group2 etc. Names and number of columns in the dataframe is set by user to define grouping variables in RMark dataframe

Value

list with first component the matrix of encounter histories, second components the vector of number of individuals with corresponding histories and, if relevant, third component vector/matrix with group(s)

Author

Olivier Gimenez <olivier.gimenez@cefe.cnrs.fr>

Examples

# read in Dipper dataset
dipper = system.file("extdata", "ed.inp", package = "R2ucare")
read_inp(dipper,group.df=data.frame(sex=c('Male','Female')))

#> $encounter_histories
#>        [,1] [,2] [,3] [,4] [,5] [,6] [,7]
#>   [1,]    1    1    1    1    1    1    0
#>   [2,]    1    1    1    1    0    0    0
#>   [3,]    1    1    0    0    0    0    0
#>   [4,]    1    1    0    0    0    0    0
#>   [5,]    1    1    0    0    0    0    0
#>   [6,]    1    1    0    0    0    0    0
#>   [7,]    1    0    1    0    0    0    0
#>   [8,]    1    0    0    0    0    0    0
#>   [9,]    1    0    0    0    0    0    0
#>  [10,]    1    0    0    0    0    0    0
#>  [11,]    1    0    0    0    0    0    0
#>  [12,]    1    0    0    0    0    0    0
#>  [13,]    0    1    1    1    1    0    0
#>  [14,]    0    1    1    1    0    0    0
#>  [15,]    0    1    1    0    0    0    0
#>  [16,]    0    1    1    0    0    0    0
#>  [17,]    0    1    1    0    0    0    0
#>  [18,]    0    1    1    0    0    0    0
#>  [19,]    0    1    1    0    0    0    0
#>  [20,]    0    1    1    0    0    0    0
#>  [21,]    0    1    1    0    0    0    0
#>  [22,]    0    1    0    0    0    0    0
#>  [23,]    0    1    0    0    0    0    0
#>  [24,]    0    1    0    0    0    0    0
#>  [25,]    0    1    0    0    0    0    0
#>  [26,]    0    1    0    0    0    0    0
#>  [27,]    0    1    0    0    0    0    0
#>  [28,]    0    1    0    0    0    0    0
#>  [29,]    0    1    0    0    0    0    0
#>  [30,]    0    1    0    0    0    0    0
#>  [31,]    0    1    0    0    0    0    0
#>  [32,]    0    1    0    0    0    0    0
#>  [33,]    0    0    1    1    1    1    0
#>  [34,]    0    0    1    1    1    0    0
#>  [35,]    0    0    1    1    1    0    0
#>  [36,]    0    0    1    1    1    0    0
#>  [37,]    0    0    1    1    1    0    0
#>  [38,]    0    0    1    1    0    0    0
#>  [39,]    0    0    1    1    0    0    0
#>  [40,]    0    0    1    1    0    0    0
#>  [41,]    0    0    1    1    0    0    0
#>  [42,]    0    0    1    1    0    0    0
#>  [43,]    0    0    1    1    0    0    0
#>  [44,]    0    0    1    1    0    0    0
#>  [45,]    0    0    1    1    0    0    0
#>  [46,]    0    0    1    0    1    1    0
#>  [47,]    0    0    1    0    0    0    0
#>  [48,]    0    0    1    0    0    0    0
#>  [49,]    0    0    1    0    0    0    0
#>  [50,]    0    0    1    0    0    0    0
#>  [51,]    0    0    1    0    0    0    0
#>  [52,]    0    0    1    0    0    0    0
#>  [53,]    0    0    1    0    0    0    0
#>  [54,]    0    0    1    0    0    0    0
#>  [55,]    0    0    1    0    0    0    0
#>  [56,]    0    0    1    0    0    0    0
#>  [57,]    0    0    1    0    0    0    0
#>  [58,]    0    0    0    1    1    1    1
#>  [59,]    0    0    0    1    1    1    1
#>  [60,]    0    0    0    1    1    1    1
#>  [61,]    0    0    0    1    1    1    1
#>  [62,]    0    0    0    1    1    1    1
#>  [63,]    0    0    0    1    1    1    1
#>  [64,]    0    0    0    1    1    1    0
#>  [65,]    0    0    0    1    1    1    0
#>  [66,]    0    0    0    1    1    1    0
#>  [67,]    0    0    0    1    1    0    0
#>  [68,]    0    0    0    1    1    0    0
#>  [69,]    0    0    0    1    1    0    0
#>  [70,]    0    0    0    1    1    0    0
#>  [71,]    0    0    0    1    1    0    0
#>  [72,]    0    0    0    1    1    0    0
#>  [73,]    0    0    0    1    0    0    1
#>  [74,]    0    0    0    1    0    0    0
#>  [75,]    0    0    0    1    0    0    0
#>  [76,]    0    0    0    1    0    0    0
#>  [77,]    0    0    0    1    0    0    0
#>  [78,]    0    0    0    1    0    0    0
#>  [79,]    0    0    0    1    0    0    0
#>  [80,]    0    0    0    0    1    1    1
#>  [81,]    0    0    0    0    1    1    1
#>  [82,]    0    0    0    0    1    1    1
#>  [83,]    0    0    0    0    1    1    1
#>  [84,]    0    0    0    0    1    1    1
#>  [85,]    0    0    0    0    1    1    1
#>  [86,]    0    0    0    0    1    1    1
#>  [87,]    0    0    0    0    1    1    1
#>  [88,]    0    0    0    0    1    1    1
#>  [89,]    0    0    0    0    1    1    1
#>  [90,]    0    0    0    0    1    1    0
#>  [91,]    0    0    0    0    1    1    0
#>  [92,]    0    0    0    0    1    1    0
#>  [93,]    0    0    0    0    1    0    0
#>  [94,]    0    0    0    0    1    0    0
#>  [95,]    0    0    0    0    1    0    0
#>  [96,]    0    0    0    0    1    0    0
#>  [97,]    0    0    0    0    1    0    0
#>  [98,]    0    0    0    0    1    0    0
#>  [99,]    0    0    0    0    1    0    0
#> [100,]    0    0    0    0    1    0    0
#> [101,]    0    0    0    0    1    0    0
#> [102,]    0    0    0    0    0    1    1
#> [103,]    0    0    0    0    0    1    1
#> [104,]    0    0    0    0    0    1    1
#> [105,]    0    0    0    0    0    1    1
#> [106,]    0    0    0    0    0    1    1
#> [107,]    0    0    0    0    0    1    1
#> [108,]    0    0    0    0    0    1    1
#> [109,]    0    0    0    0    0    1    1
#> [110,]    0    0    0    0    0    1    1
#> [111,]    0    0    0    0    0    1    1
#> [112,]    0    0    0    0    0    1    1
#> [113,]    0    0    0    0    0    1    1
#> [114,]    0    0    0    0    0    1    0
#> [115,]    0    0    0    0    0    1    0
#> [116,]    0    0    0    0    0    1    0
#> [117,]    0    0    0    0    0    1    0
#> [118,]    0    0    0    0    0    1    0
#> [119,]    0    0    0    0    0    1    0
#> [120,]    0    0    0    0    0    1    0
#> [121,]    0    0    0    0    0    1    0
#> [122,]    0    0    0    0    0    1    0
#> [123,]    0    0    0    0    0    1    0
#> [124,]    0    0    0    0    0    1    0
#> [125,]    0    0    0    0    0    0    1
#> [126,]    0    0    0    0    0    0    1
#> [127,]    0    0    0    0    0    0    1
#> [128,]    0    0    0    0    0    0    1
#> [129,]    0    0    0    0    0    0    1
#> [130,]    0    0    0    0    0    0    1
#> [131,]    0    0    0    0    0    0    1
#> [132,]    0    0    0    0    0    0    1
#> [133,]    0    0    0    0    0    0    1
#> [134,]    0    0    0    0    0    0    1
#> [135,]    0    0    0    0    0    0    1
#> [136,]    0    0    0    0    0    0    1
#> [137,]    0    0    0    0    0    0    1
#> [138,]    0    0    0    0    0    0    1
#> [139,]    0    0    0    0    0    0    1
#> [140,]    0    0    0    0    0    0    1
#> [141,]    0    0    0    0    0    0    1
#> [142,]    1    1    1    1    1    0    0
#> [143,]    1    1    1    1    0    0    0
#> [144,]    1    1    0    1    1    1    0
#> [145,]    1    1    0    0    0    0    0
#> [146,]    1    1    0    0    0    0    0
#> [147,]    1    0    1    0    0    0    0
#> [148,]    1    0    0    0    0    0    0
#> [149,]    1    0    0    0    0    0    0
#> [150,]    1    0    0    0    0    0    0
#> [151,]    1    0    0    0    0    0    0
#> [152,]    0    1    1    1    1    1    1
#> [153,]    0    1    1    1    1    1    1
#> [154,]    0    1    1    1    1    1    0
#> [155,]    0    1    1    1    1    0    0
#> [156,]    0    1    1    1    1    0    0
#> [157,]    0    1    1    1    0    0    0
#> [158,]    0    1    1    0    1    1    0
#> [159,]    0    1    1    0    0    0    0
#> [160,]    0    1    1    0    0    0    0
#> [161,]    0    1    1    0    0    0    0
#> [162,]    0    1    1    0    0    0    0
#> [163,]    0    1    0    0    0    0    0
#> [164,]    0    1    0    0    0    0    0
#> [165,]    0    1    0    0    0    0    0
#> [166,]    0    1    0    0    0    0    0
#> [167,]    0    1    0    0    0    0    0
#> [168,]    0    1    0    0    0    0    0
#> [169,]    0    1    0    0    0    0    0
#> [170,]    0    1    0    0    0    0    0
#> [171,]    0    1    0    0    0    0    0
#> [172,]    0    1    0    0    0    0    0
#> [173,]    0    1    0    0    0    0    0
#> [174,]    0    1    0    0    0    0    0
#> [175,]    0    1    0    0    0    0    0
#> [176,]    0    1    0    0    0    0    0
#> [177,]    0    1    0    0    0    0    0
#> [178,]    0    1    0    0    0    0    0
#> [179,]    0    1    0    0    0    0    0
#> [180,]    0    1    0    0    0    0    0
#> [181,]    0    0    1    1    1    1    1
#> [182,]    0    0    1    1    1    1    1
#> [183,]    0    0    1    1    1    1    0
#> [184,]    0    0    1    1    1    0    0
#> [185,]    0    0    1    1    1    0    0
#> [186,]    0    0    1    1    0    0    0
#> [187,]    0    0    1    1    0    0    0
#> [188,]    0    0    1    1    0    0    0
#> [189,]    0    0    1    1    0    0    0
#> [190,]    0    0    1    0    0    0    0
#> [191,]    0    0    1    0    0    0    0
#> [192,]    0    0    1    0    0    0    0
#> [193,]    0    0    1    0    0    0    0
#> [194,]    0    0    1    0    0    0    0
#> [195,]    0    0    1    0    0    0    0
#> [196,]    0    0    1    0    0    0    0
#> [197,]    0    0    1    0    0    0    0
#> [198,]    0    0    1    0    0    0    0
#> [199,]    0    0    1    0    0    0    0
#> [200,]    0    0    1    0    0    0    0
#> [201,]    0    0    1    0    0    0    0
#> [202,]    0    0    1    0    0    0    0
#> [203,]    0    0    1    0    0    0    0
#> [204,]    0    0    1    0    0    0    0
#> [205,]    0    0    1    0    0    0    0
#> [206,]    0    0    1    0    0    0    0
#> [207,]    0    0    1    0    0    0    0
#> [208,]    0    0    0    1    1    1    1
#> [209,]    0    0    0    1    1    1    1
#> [210,]    0    0    0    1    1    1    0
#> [211,]    0    0    0    1    1    1    0
#> [212,]    0    0    0    1    1    1    0
#> [213,]    0    0    0    1    1    1    0
#> [214,]    0    0    0    1    1    0    0
#> [215,]    0    0    0    1    1    0    0
#> [216,]    0    0    0    1    1    0    0
#> [217,]    0    0    0    1    1    0    0
#> [218,]    0    0    0    1    1    0    0
#> [219,]    0    0    0    1    0    1    1
#> [220,]    0    0    0    1    0    0    1
#> [221,]    0    0    0    1    0    0    0
#> [222,]    0    0    0    1    0    0    0
#> [223,]    0    0    0    1    0    0    0
#> [224,]    0    0    0    1    0    0    0
#> [225,]    0    0    0    1    0    0    0
#> [226,]    0    0    0    1    0    0    0
#> [227,]    0    0    0    1    0    0    0
#> [228,]    0    0    0    1    0    0    0
#> [229,]    0    0    0    1    0    0    0
#> [230,]    0    0    0    1    0    0    0
#> [231,]    0    0    0    0    1    1    1
#> [232,]    0    0    0    0    1    1    1
#> [233,]    0    0    0    0    1    1    1
#> [234,]    0    0    0    0    1    1    1
#> [235,]    0    0    0    0    1    1    1
#> [236,]    0    0    0    0    1    1    1
#> [237,]    0    0    0    0    1    1    0
#> [238,]    0    0    0    0    1    1    0
#> [239,]    0    0    0    0    1    1    0
#> [240,]    0    0    0    0    1    1    0
#> [241,]    0    0    0    0    1    1    0
#> [242,]    0    0    0    0    1    1    0
#> [243,]    0    0    0    0    1    0    0
#> [244,]    0    0    0    0    1    0    0
#> [245,]    0    0    0    0    1    0    0
#> [246,]    0    0    0    0    1    0    0
#> [247,]    0    0    0    0    1    0    0
#> [248,]    0    0    0    0    1    0    0
#> [249,]    0    0    0    0    1    0    0
#> [250,]    0    0    0    0    0    1    1
#> [251,]    0    0    0    0    0    1    1
#> [252,]    0    0    0    0    0    1    1
#> [253,]    0    0    0    0    0    1    1
#> [254,]    0    0    0    0    0    1    1
#> [255,]    0    0    0    0    0    1    1
#> [256,]    0    0    0    0    0    1    1
#> [257,]    0    0    0    0    0    1    1
#> [258,]    0    0    0    0    0    1    1
#> [259,]    0    0    0    0    0    1    1
#> [260,]    0    0    0    0    0    1    1
#> [261,]    0    0    0    0    0    1    0
#> [262,]    0    0    0    0    0    1    0
#> [263,]    0    0    0    0    0    1    0
#> [264,]    0    0    0    0    0    1    0
#> [265,]    0    0    0    0    0    1    0
#> [266,]    0    0    0    0    0    1    0
#> [267,]    0    0    0    0    0    1    0
#> [268,]    0    0    0    0    0    1    0
#> [269,]    0    0    0    0    0    1    0
#> [270,]    0    0    0    0    0    1    0
#> [271,]    0    0    0    0    0    1    0
#> [272,]    0    0    0    0    0    1    0
#> [273,]    0    0    0    0    0    0    1
#> [274,]    0    0    0    0    0    0    1
#> [275,]    0    0    0    0    0    0    1
#> [276,]    0    0    0    0    0    0    1
#> [277,]    0    0    0    0    0    0    1
#> [278,]    0    0    0    0    0    0    1
#> [279,]    0    0    0    0    0    0    1
#> [280,]    0    0    0    0    0    0    1
#> [281,]    0    0    0    0    0    0    1
#> [282,]    0    0    0    0    0    0    1
#> [283,]    0    0    0    0    0    0    1
#> [284,]    0    0    0    0    0    0    1
#> [285,]    0    0    0    0    0    0    1
#> [286,]    0    0    0    0    0    0    1
#> [287,]    0    0    0    0    0    0    1
#> [288,]    0    0    0    0    0    0    1
#> [289,]    0    0    0    0    0    0    1
#> [290,]    0    0    0    0    0    0    1
#> [291,]    0    0    0    0    0    0    1
#> [292,]    0    0    0    0    0    0    1
#> [293,]    0    0    0    0    0    0    1
#> [294,]    0    0    0    0    0    0    1
#> 
#> $sample_size
#>   [1] 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1
#>  [38] 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1
#>  [75] 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1
#> [112] 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1
#> [149] 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1
#> [186] 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1
#> [223] 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1
#> [260] 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1
#> 
#> $groups
#>   [1] "Male"   "Male"   "Male"   "Male"   "Male"   "Male"   "Male"   "Male"  
#>   [9] "Male"   "Male"   "Male"   "Male"   "Male"   "Male"   "Male"   "Male"  
#>  [17] "Male"   "Male"   "Male"   "Male"   "Male"   "Male"   "Male"   "Male"  
#>  [25] "Male"   "Male"   "Male"   "Male"   "Male"   "Male"   "Male"   "Male"  
#>  [33] "Male"   "Male"   "Male"   "Male"   "Male"   "Male"   "Male"   "Male"  
#>  [41] "Male"   "Male"   "Male"   "Male"   "Male"   "Male"   "Male"   "Male"  
#>  [49] "Male"   "Male"   "Male"   "Male"   "Male"   "Male"   "Male"   "Male"  
#>  [57] "Male"   "Male"   "Male"   "Male"   "Male"   "Male"   "Male"   "Male"  
#>  [65] "Male"   "Male"   "Male"   "Male"   "Male"   "Male"   "Male"   "Male"  
#>  [73] "Male"   "Male"   "Male"   "Male"   "Male"   "Male"   "Male"   "Male"  
#>  [81] "Male"   "Male"   "Male"   "Male"   "Male"   "Male"   "Male"   "Male"  
#>  [89] "Male"   "Male"   "Male"   "Male"   "Male"   "Male"   "Male"   "Male"  
#>  [97] "Male"   "Male"   "Male"   "Male"   "Male"   "Male"   "Male"   "Male"  
#> [105] "Male"   "Male"   "Male"   "Male"   "Male"   "Male"   "Male"   "Male"  
#> [113] "Male"   "Male"   "Male"   "Male"   "Male"   "Male"   "Male"   "Male"  
#> [121] "Male"   "Male"   "Male"   "Male"   "Male"   "Male"   "Male"   "Male"  
#> [129] "Male"   "Male"   "Male"   "Male"   "Male"   "Male"   "Male"   "Male"  
#> [137] "Male"   "Male"   "Male"   "Male"   "Male"   "Female" "Female" "Female"
#> [145] "Female" "Female" "Female" "Female" "Female" "Female" "Female" "Female"
#> [153] "Female" "Female" "Female" "Female" "Female" "Female" "Female" "Female"
#> [161] "Female" "Female" "Female" "Female" "Female" "Female" "Female" "Female"
#> [169] "Female" "Female" "Female" "Female" "Female" "Female" "Female" "Female"
#> [177] "Female" "Female" "Female" "Female" "Female" "Female" "Female" "Female"
#> [185] "Female" "Female" "Female" "Female" "Female" "Female" "Female" "Female"
#> [193] "Female" "Female" "Female" "Female" "Female" "Female" "Female" "Female"
#> [201] "Female" "Female" "Female" "Female" "Female" "Female" "Female" "Female"
#> [209] "Female" "Female" "Female" "Female" "Female" "Female" "Female" "Female"
#> [217] "Female" "Female" "Female" "Female" "Female" "Female" "Female" "Female"
#> [225] "Female" "Female" "Female" "Female" "Female" "Female" "Female" "Female"
#> [233] "Female" "Female" "Female" "Female" "Female" "Female" "Female" "Female"
#> [241] "Female" "Female" "Female" "Female" "Female" "Female" "Female" "Female"
#> [249] "Female" "Female" "Female" "Female" "Female" "Female" "Female" "Female"
#> [257] "Female" "Female" "Female" "Female" "Female" "Female" "Female" "Female"
#> [265] "Female" "Female" "Female" "Female" "Female" "Female" "Female" "Female"
#> [273] "Female" "Female" "Female" "Female" "Female" "Female" "Female" "Female"
#> [281] "Female" "Female" "Female" "Female" "Female" "Female" "Female" "Female"
#> [289] "Female" "Female" "Female" "Female" "Female" "Female"
#> 

# read in Geese dataset
geese = system.file("extdata", "geese.inp", package = "R2ucare")
read_inp(geese)

#> $encounter_histories
#>        [,1] [,2] [,3] [,4] [,5] [,6]
#>   [1,]    0    0    0    0    0    1
#>   [2,]    0    0    0    0    0    2
#>   [3,]    0    0    0    0    0    3
#>   [4,]    0    0    0    0    1    0
#>   [5,]    0    0    0    0    1    1
#>   [6,]    0    0    0    0    1    2
#>   [7,]    0    0    0    0    2    0
#>   [8,]    0    0    0    0    2    1
#>   [9,]    0    0    0    0    2    2
#>  [10,]    0    0    0    0    2    3
#>  [11,]    0    0    0    0    3    0
#>  [12,]    0    0    0    0    3    1
#>  [13,]    0    0    0    0    3    2
#>  [14,]    0    0    0    0    3    3
#>  [15,]    0    0    0    1    0    0
#>  [16,]    0    0    0    1    0    1
#>  [17,]    0    0    0    1    0    2
#>  [18,]    0    0    0    1    1    0
#>  [19,]    0    0    0    1    1    1
#>  [20,]    0    0    0    1    1    2
#>  [21,]    0    0    0    1    2    0
#>  [22,]    0    0    0    1    2    1
#>  [23,]    0    0    0    1    2    2
#>  [24,]    0    0    0    1    2    3
#>  [25,]    0    0    0    1    3    0
#>  [26,]    0    0    0    1    3    2
#>  [27,]    0    0    0    2    0    0
#>  [28,]    0    0    0    2    0    1
#>  [29,]    0    0    0    2    0    2
#>  [30,]    0    0    0    2    0    3
#>  [31,]    0    0    0    2    1    0
#>  [32,]    0    0    0    2    1    1
#>  [33,]    0    0    0    2    1    2
#>  [34,]    0    0    0    2    2    0
#>  [35,]    0    0    0    2    2    1
#>  [36,]    0    0    0    2    2    2
#>  [37,]    0    0    0    2    2    3
#>  [38,]    0    0    0    2    3    0
#>  [39,]    0    0    0    2    3    1
#>  [40,]    0    0    0    2    3    2
#>  [41,]    0    0    0    2    3    3
#>  [42,]    0    0    0    3    0    0
#>  [43,]    0    0    0    3    0    1
#>  [44,]    0    0    0    3    0    2
#>  [45,]    0    0    0    3    0    3
#>  [46,]    0    0    0    3    1    0
#>  [47,]    0    0    0    3    1    1
#>  [48,]    0    0    0    3    1    3
#>  [49,]    0    0    0    3    2    0
#>  [50,]    0    0    0    3    2    1
#>  [51,]    0    0    0    3    2    2
#>  [52,]    0    0    0    3    2    3
#>  [53,]    0    0    0    3    3    0
#>  [54,]    0    0    0    3    3    1
#>  [55,]    0    0    0    3    3    2
#>  [56,]    0    0    0    3    3    3
#>  [57,]    0    0    1    0    0    0
#>  [58,]    0    0    1    0    0    1
#>  [59,]    0    0    1    0    0    2
#>  [60,]    0    0    1    0    0    3
#>  [61,]    0    0    1    0    1    0
#>  [62,]    0    0    1    0    1    1
#>  [63,]    0    0    1    0    1    2
#>  [64,]    0    0    1    0    2    0
#>  [65,]    0    0    1    0    2    1
#>  [66,]    0    0    1    0    2    2
#>  [67,]    0    0    1    0    3    1
#>  [68,]    0    0    1    0    3    2
#>  [69,]    0    0    1    0    3    3
#>  [70,]    0    0    1    1    0    0
#>  [71,]    0    0    1    1    0    1
#>  [72,]    0    0    1    1    0    2
#>  [73,]    0    0    1    1    1    0
#>  [74,]    0    0    1    1    1    1
#>  [75,]    0    0    1    1    1    2
#>  [76,]    0    0    1    1    2    0
#>  [77,]    0    0    1    1    2    1
#>  [78,]    0    0    1    1    2    2
#>  [79,]    0    0    1    2    0    0
#>  [80,]    0    0    1    2    0    1
#>  [81,]    0    0    1    2    0    2
#>  [82,]    0    0    1    2    1    0
#>  [83,]    0    0    1    2    1    1
#>  [84,]    0    0    1    2    1    2
#>  [85,]    0    0    1    2    2    0
#>  [86,]    0    0    1    2    2    1
#>  [87,]    0    0    1    2    2    2
#>  [88,]    0    0    1    2    3    0
#>  [89,]    0    0    1    3    0    0
#>  [90,]    0    0    1    3    1    0
#>  [91,]    0    0    2    0    0    0
#>  [92,]    0    0    2    0    0    1
#>  [93,]    0    0    2    0    0    2
#>  [94,]    0    0    2    0    0    3
#>  [95,]    0    0    2    0    1    0
#>  [96,]    0    0    2    0    1    1
#>  [97,]    0    0    2    0    1    2
#>  [98,]    0    0    2    0    2    0
#>  [99,]    0    0    2    0    2    1
#> [100,]    0    0    2    0    2    2
#> [101,]    0    0    2    0    2    3
#> [102,]    0    0    2    0    3    0
#> [103,]    0    0    2    0    3    2
#> [104,]    0    0    2    1    0    0
#> [105,]    0    0    2    1    0    1
#> [106,]    0    0    2    1    0    2
#> [107,]    0    0    2    1    1    0
#> [108,]    0    0    2    1    1    1
#> [109,]    0    0    2    1    1    2
#> [110,]    0    0    2    1    2    0
#> [111,]    0    0    2    1    2    1
#> [112,]    0    0    2    1    2    2
#> [113,]    0    0    2    1    3    1
#> [114,]    0    0    2    2    0    0
#> [115,]    0    0    2    2    0    1
#> [116,]    0    0    2    2    0    2
#> [117,]    0    0    2    2    1    0
#> [118,]    0    0    2    2    1    1
#> [119,]    0    0    2    2    1    2
#> [120,]    0    0    2    2    2    0
#> [121,]    0    0    2    2    2    1
#> [122,]    0    0    2    2    2    2
#> [123,]    0    0    2    2    3    0
#> [124,]    0    0    2    2    3    2
#> [125,]    0    0    2    3    0    0
#> [126,]    0    0    2    3    2    0
#> [127,]    0    0    2    3    2    1
#> [128,]    0    0    2    3    3    0
#> [129,]    0    0    3    0    0    0
#> [130,]    0    0    3    0    0    1
#> [131,]    0    0    3    0    0    2
#> [132,]    0    0    3    0    0    3
#> [133,]    0    0    3    0    1    0
#> [134,]    0    0    3    0    1    2
#> [135,]    0    0    3    0    2    0
#> [136,]    0    0    3    0    2    2
#> [137,]    0    0    3    0    2    3
#> [138,]    0    0    3    0    3    0
#> [139,]    0    0    3    0    3    2
#> [140,]    0    0    3    0    3    3
#> [141,]    0    0    3    1    0    0
#> [142,]    0    0    3    2    0    0
#> [143,]    0    0    3    2    0    2
#> [144,]    0    0    3    2    2    0
#> [145,]    0    0    3    2    2    1
#> [146,]    0    0    3    2    2    2
#> [147,]    0    0    3    2    3    0
#> [148,]    0    0    3    2    3    3
#> [149,]    0    0    3    3    0    0
#> [150,]    0    0    3    3    0    1
#> [151,]    0    0    3    3    0    2
#> [152,]    0    0    3    3    0    3
#> [153,]    0    0    3    3    2    0
#> [154,]    0    0    3    3    3    0
#> [155,]    0    0    3    3    3    2
#> [156,]    0    0    3    3    3    3
#> [157,]    0    1    0    0    0    0
#> [158,]    0    1    0    0    0    1
#> [159,]    0    1    0    0    0    2
#> [160,]    0    1    0    0    1    0
#> [161,]    0    1    0    0    1    1
#> [162,]    0    1    0    0    1    2
#> [163,]    0    1    0    0    2    0
#> [164,]    0    1    0    0    2    1
#> [165,]    0    1    0    0    2    2
#> [166,]    0    1    0    0    3    0
#> [167,]    0    1    0    1    0    0
#> [168,]    0    1    0    1    0    1
#> [169,]    0    1    0    1    0    2
#> [170,]    0    1    0    1    1    0
#> [171,]    0    1    0    1    1    1
#> [172,]    0    1    0    1    2    0
#> [173,]    0    1    0    2    0    0
#> [174,]    0    1    0    2    0    1
#> [175,]    0    1    0    2    0    2
#> [176,]    0    1    0    2    1    0
#> [177,]    0    1    0    2    1    2
#> [178,]    0    1    0    2    2    0
#> [179,]    0    1    0    2    2    1
#> [180,]    0    1    0    2    2    2
#> [181,]    0    1    0    3    0    0
#> [182,]    0    1    0    3    3    0
#> [183,]    0    1    1    0    0    0
#> [184,]    0    1    1    0    0    1
#> [185,]    0    1    1    0    0    2
#> [186,]    0    1    1    0    1    0
#> [187,]    0    1    1    0    1    1
#> [188,]    0    1    1    0    1    2
#> [189,]    0    1    1    0    2    0
#> [190,]    0    1    1    0    2    2
#> [191,]    0    1    1    1    0    0
#> [192,]    0    1    1    1    0    1
#> [193,]    0    1    1    1    0    2
#> [194,]    0    1    1    1    1    0
#> [195,]    0    1    1    1    1    1
#> [196,]    0    1    1    1    1    2
#> [197,]    0    1    1    1    2    0
#> [198,]    0    1    1    2    0    0
#> [199,]    0    1    1    2    1    0
#> [200,]    0    1    1    2    2    0
#> [201,]    0    1    1    2    2    1
#> [202,]    0    1    2    0    0    0
#> [203,]    0    1    2    0    0    1
#> [204,]    0    1    2    0    0    2
#> [205,]    0    1    2    0    1    0
#> [206,]    0    1    2    0    1    1
#> [207,]    0    1    2    0    2    0
#> [208,]    0    1    2    0    2    2
#> [209,]    0    1    2    1    0    0
#> [210,]    0    1    2    1    1    0
#> [211,]    0    1    2    1    1    1
#> [212,]    0    1    2    2    0    0
#> [213,]    0    1    2    2    0    2
#> [214,]    0    1    2    2    2    0
#> [215,]    0    1    2    2    2    1
#> [216,]    0    1    2    3    0    0
#> [217,]    0    2    0    0    0    0
#> [218,]    0    2    0    0    0    1
#> [219,]    0    2    0    0    0    2
#> [220,]    0    2    0    0    0    3
#> [221,]    0    2    0    0    1    0
#> [222,]    0    2    0    0    1    1
#> [223,]    0    2    0    0    1    2
#> [224,]    0    2    0    0    2    0
#> [225,]    0    2    0    0    2    1
#> [226,]    0    2    0    0    2    2
#> [227,]    0    2    0    0    3    0
#> [228,]    0    2    0    0    3    2
#> [229,]    0    2    0    1    0    0
#> [230,]    0    2    0    1    0    1
#> [231,]    0    2    0    1    0    2
#> [232,]    0    2    0    1    1    0
#> [233,]    0    2    0    1    1    1
#> [234,]    0    2    0    1    1    2
#> [235,]    0    2    0    1    2    0
#> [236,]    0    2    0    1    2    1
#> [237,]    0    2    0    1    2    2
#> [238,]    0    2    0    2    0    0
#> [239,]    0    2    0    2    0    1
#> [240,]    0    2    0    2    0    2
#> [241,]    0    2    0    2    0    3
#> [242,]    0    2    0    2    1    0
#> [243,]    0    2    0    2    1    2
#> [244,]    0    2    0    2    2    0
#> [245,]    0    2    0    2    2    2
#> [246,]    0    2    0    2    3    0
#> [247,]    0    2    0    3    0    0
#> [248,]    0    2    0    3    0    3
#> [249,]    0    2    0    3    1    0
#> [250,]    0    2    1    0    0    0
#> [251,]    0    2    1    0    0    2
#> [252,]    0    2    1    0    1    0
#> [253,]    0    2    1    0    1    1
#> [254,]    0    2    1    0    1    2
#> [255,]    0    2    1    0    2    0
#> [256,]    0    2    1    0    2    1
#> [257,]    0    2    1    0    2    2
#> [258,]    0    2    1    1    0    0
#> [259,]    0    2    1    1    1    0
#> [260,]    0    2    1    1    1    1
#> [261,]    0    2    1    1    1    2
#> [262,]    0    2    1    1    2    0
#> [263,]    0    2    1    1    2    1
#> [264,]    0    2    1    2    0    0
#> [265,]    0    2    1    2    0    2
#> [266,]    0    2    1    2    1    0
#> [267,]    0    2    1    2    2    0
#> [268,]    0    2    1    2    2    2
#> [269,]    0    2    1    3    0    2
#> [270,]    0    2    1    3    1    0
#> [271,]    0    2    2    0    0    0
#> [272,]    0    2    2    0    0    1
#> [273,]    0    2    2    0    0    2
#> [274,]    0    2    2    0    1    0
#> [275,]    0    2    2    0    1    1
#> [276,]    0    2    2    0    1    2
#> [277,]    0    2    2    0    2    0
#> [278,]    0    2    2    0    2    1
#> [279,]    0    2    2    0    2    2
#> [280,]    0    2    2    0    3    0
#> [281,]    0    2    2    0    3    3
#> [282,]    0    2    2    1    0    0
#> [283,]    0    2    2    1    0    2
#> [284,]    0    2    2    1    1    0
#> [285,]    0    2    2    1    2    0
#> [286,]    0    2    2    2    0    0
#> [287,]    0    2    2    2    0    1
#> [288,]    0    2    2    2    0    2
#> [289,]    0    2    2    2    1    0
#> [290,]    0    2    2    2    1    1
#> [291,]    0    2    2    2    2    0
#> [292,]    0    2    2    2    2    2
#> [293,]    0    2    2    3    0    0
#> [294,]    0    2    2    3    0    3
#> [295,]    0    2    3    0    0    0
#> [296,]    0    2    3    0    0    2
#> [297,]    0    2    3    0    0    3
#> [298,]    0    2    3    0    2    0
#> [299,]    0    2    3    3    0    0
#> [300,]    0    2    3    3    3    0
#> [301,]    0    3    0    0    0    0
#> [302,]    0    3    0    0    0    1
#> [303,]    0    3    0    0    0    2
#> [304,]    0    3    0    0    0    3
#> [305,]    0    3    0    0    1    0
#> [306,]    0    3    0    0    1    1
#> [307,]    0    3    0    0    1    2
#> [308,]    0    3    0    0    2    0
#> [309,]    0    3    0    0    2    2
#> [310,]    0    3    0    0    2    3
#> [311,]    0    3    0    0    3    0
#> [312,]    0    3    0    0    3    2
#> [313,]    0    3    0    0    3    3
#> [314,]    0    3    0    1    0    0
#> [315,]    0    3    0    1    0    1
#> [316,]    0    3    0    1    2    0
#> [317,]    0    3    0    2    0    0
#> [318,]    0    3    0    2    0    2
#> [319,]    0    3    0    2    0    3
#> [320,]    0    3    0    2    2    0
#> [321,]    0    3    0    2    2    2
#> [322,]    0    3    0    2    3    0
#> [323,]    0    3    0    3    0    0
#> [324,]    0    3    0    3    0    1
#> [325,]    0    3    0    3    0    3
#> [326,]    0    3    0    3    2    3
#> [327,]    0    3    0    3    3    0
#> [328,]    0    3    0    3    3    3
#> [329,]    0    3    1    0    0    0
#> [330,]    0    3    1    0    0    2
#> [331,]    0    3    1    1    1    0
#> [332,]    0    3    1    2    0    0
#> [333,]    0    3    1    2    1    0
#> [334,]    0    3    1    3    0    1
#> [335,]    0    3    2    0    0    0
#> [336,]    0    3    2    0    0    2
#> [337,]    0    3    2    0    0    3
#> [338,]    0    3    2    0    1    0
#> [339,]    0    3    2    0    2    0
#> [340,]    0    3    2    1    1    0
#> [341,]    0    3    2    1    2    0
#> [342,]    0    3    2    2    0    0
#> [343,]    0    3    2    2    0    2
#> [344,]    0    3    2    2    1    0
#> [345,]    0    3    2    2    2    0
#> [346,]    0    3    2    2    2    2
#> [347,]    0    3    2    3    0    0
#> [348,]    0    3    2    3    3    0
#> [349,]    0    3    2    3    3    2
#> [350,]    0    3    3    0    0    0
#> [351,]    0    3    3    0    0    2
#> [352,]    0    3    3    0    0    3
#> [353,]    0    3    3    0    1    0
#> [354,]    0    3    3    0    2    0
#> [355,]    0    3    3    0    3    3
#> [356,]    0    3    3    1    0    2
#> [357,]    0    3    3    2    0    0
#> [358,]    0    3    3    2    0    2
#> [359,]    0    3    3    2    2    0
#> [360,]    0    3    3    2    3    0
#> [361,]    0    3    3    2    3    3
#> [362,]    0    3    3    3    0    0
#> [363,]    0    3    3    3    0    3
#> [364,]    0    3    3    3    2    0
#> [365,]    0    3    3    3    3    0
#> [366,]    0    3    3    3    3    3
#> [367,]    1    0    0    0    0    0
#> [368,]    1    0    0    0    0    1
#> [369,]    1    0    0    0    0    2
#> [370,]    1    0    0    0    1    0
#> [371,]    1    0    0    0    2    0
#> [372,]    1    0    0    0    2    2
#> [373,]    1    0    0    0    3    0
#> [374,]    1    0    0    1    0    0
#> [375,]    1    0    0    1    0    1
#> [376,]    1    0    0    1    1    0
#> [377,]    1    0    0    1    1    1
#> [378,]    1    0    0    1    2    0
#> [379,]    1    0    0    2    0    0
#> [380,]    1    0    0    2    0    2
#> [381,]    1    0    1    0    0    0
#> [382,]    1    0    1    0    0    2
#> [383,]    1    0    1    0    1    0
#> [384,]    1    0    1    0    2    0
#> [385,]    1    0    1    1    0    0
#> [386,]    1    0    1    1    0    1
#> [387,]    1    0    1    1    0    2
#> [388,]    1    0    1    2    1    1
#> [389,]    1    0    2    0    0    0
#> [390,]    1    0    2    0    0    1
#> [391,]    1    0    2    0    1    0
#> [392,]    1    0    2    0    2    2
#> [393,]    1    0    2    1    0    1
#> [394,]    1    0    2    1    2    0
#> [395,]    1    0    2    2    0    0
#> [396,]    1    0    2    2    2    0
#> [397,]    1    1    0    0    0    0
#> [398,]    1    1    0    0    0    1
#> [399,]    1    1    0    0    1    0
#> [400,]    1    1    0    0    2    0
#> [401,]    1    1    0    1    0    0
#> [402,]    1    1    0    1    0    1
#> [403,]    1    1    0    1    1    0
#> [404,]    1    1    0    1    1    1
#> [405,]    1    1    0    2    0    0
#> [406,]    1    1    0    2    0    1
#> [407,]    1    1    1    0    0    0
#> [408,]    1    1    1    0    0    1
#> [409,]    1    1    1    0    1    0
#> [410,]    1    1    1    0    1    1
#> [411,]    1    1    1    0    1    2
#> [412,]    1    1    1    0    2    0
#> [413,]    1    1    1    1    0    0
#> [414,]    1    1    1    1    0    1
#> [415,]    1    1    1    1    1    0
#> [416,]    1    1    1    1    1    1
#> [417,]    1    1    1    1    2    0
#> [418,]    1    1    1    2    0    0
#> [419,]    1    1    1    2    0    1
#> [420,]    1    1    2    0    0    0
#> [421,]    1    2    0    0    0    0
#> [422,]    1    2    0    0    0    2
#> [423,]    1    2    0    0    2    0
#> [424,]    1    2    0    1    0    0
#> [425,]    1    2    0    2    0    1
#> [426,]    1    2    1    0    0    0
#> [427,]    1    2    1    0    0    1
#> [428,]    1    2    1    1    0    0
#> [429,]    1    2    1    1    1    0
#> [430,]    1    2    1    1    1    1
#> [431,]    1    2    1    2    2    0
#> [432,]    1    2    2    0    0    0
#> [433,]    2    0    0    0    0    0
#> [434,]    2    0    0    0    0    1
#> [435,]    2    0    0    0    0    2
#> [436,]    2    0    0    0    1    0
#> [437,]    2    0    0    0    1    1
#> [438,]    2    0    0    0    1    2
#> [439,]    2    0    0    0    2    0
#> [440,]    2    0    0    0    2    1
#> [441,]    2    0    0    0    2    2
#> [442,]    2    0    0    0    2    3
#> [443,]    2    0    0    0    3    0
#> [444,]    2    0    0    1    0    0
#> [445,]    2    0    0    1    0    2
#> [446,]    2    0    0    1    2    0
#> [447,]    2    0    0    1    3    2
#> [448,]    2    0    0    2    0    0
#> [449,]    2    0    0    2    0    1
#> [450,]    2    0    0    2    0    2
#> [451,]    2    0    0    2    2    0
#> [452,]    2    0    0    2    2    2
#> [453,]    2    0    0    3    0    0
#> [454,]    2    0    0    3    2    1
#> [455,]    2    0    0    3    3    3
#> [456,]    2    0    1    0    0    0
#> [457,]    2    0    1    0    0    1
#> [458,]    2    0    1    0    1    1
#> [459,]    2    0    1    1    0    0
#> [460,]    2    0    1    1    0    1
#> [461,]    2    0    1    1    0    2
#> [462,]    2    0    1    1    1    0
#> [463,]    2    0    1    2    0    0
#> [464,]    2    0    1    2    0    1
#> [465,]    2    0    1    2    0    2
#> [466,]    2    0    1    2    2    0
#> [467,]    2    0    2    0    0    0
#> [468,]    2    0    2    0    0    1
#> [469,]    2    0    2    0    0    2
#> [470,]    2    0    2    0    1    1
#> [471,]    2    0    2    0    1    2
#> [472,]    2    0    2    0    2    0
#> [473,]    2    0    2    0    2    2
#> [474,]    2    0    2    1    0    0
#> [475,]    2    0    2    1    1    0
#> [476,]    2    0    2    1    2    2
#> [477,]    2    0    2    2    0    0
#> [478,]    2    0    2    2    0    1
#> [479,]    2    0    2    2    0    2
#> [480,]    2    0    2    2    2    0
#> [481,]    2    0    2    2    2    2
#> [482,]    2    0    2    2    3    0
#> [483,]    2    0    2    3    0    0
#> [484,]    2    0    3    0    0    3
#> [485,]    2    0    3    3    0    0
#> [486,]    2    1    0    0    0    0
#> [487,]    2    1    0    0    0    1
#> [488,]    2    1    0    0    0    2
#> [489,]    2    1    0    0    2    0
#> [490,]    2    1    0    0    2    2
#> [491,]    2    1    0    1    0    0
#> [492,]    2    1    0    1    0    1
#> [493,]    2    1    0    1    1    0
#> [494,]    2    1    0    1    2    1
#> [495,]    2    1    0    2    0    0
#> [496,]    2    1    0    2    0    2
#> [497,]    2    1    0    2    1    0
#> [498,]    2    1    1    0    0    0
#> [499,]    2    1    1    1    0    0
#> [500,]    2    1    1    1    0    1
#> [501,]    2    1    1    1    0    2
#> [502,]    2    1    1    1    1    0
#> [503,]    2    1    1    2    0    0
#> [504,]    2    1    1    3    0    0
#> [505,]    2    1    2    0    0    0
#> [506,]    2    2    0    0    0    0
#> [507,]    2    2    0    0    0    1
#> [508,]    2    2    0    0    0    2
#> [509,]    2    2    0    0    0    3
#> [510,]    2    2    0    0    1    0
#> [511,]    2    2    0    0    1    1
#> [512,]    2    2    0    0    1    2
#> [513,]    2    2    0    0    2    0
#> [514,]    2    2    0    0    2    1
#> [515,]    2    2    0    0    2    2
#> [516,]    2    2    0    0    3    0
#> [517,]    2    2    0    1    0    0
#> [518,]    2    2    0    1    1    0
#> [519,]    2    2    0    1    1    1
#> [520,]    2    2    0    2    0    0
#> [521,]    2    2    0    2    0    1
#> [522,]    2    2    0    2    0    2
#> [523,]    2    2    0    2    1    0
#> [524,]    2    2    0    2    2    0
#> [525,]    2    2    0    2    2    1
#> [526,]    2    2    0    2    2    2
#> [527,]    2    2    0    2    3    2
#> [528,]    2    2    1    0    0    0
#> [529,]    2    2    1    0    0    1
#> [530,]    2    2    1    0    0    2
#> [531,]    2    2    1    1    0    0
#> [532,]    2    2    1    2    0    0
#> [533,]    2    2    2    0    0    0
#> [534,]    2    2    2    0    0    1
#> [535,]    2    2    2    0    0    2
#> [536,]    2    2    2    0    2    0
#> [537,]    2    2    2    0    3    0
#> [538,]    2    2    2    2    0    0
#> [539,]    2    2    2    2    0    1
#> [540,]    2    2    2    2    0    2
#> [541,]    2    2    2    2    1    0
#> [542,]    2    2    2    2    2    0
#> [543,]    2    2    2    2    2    2
#> [544,]    2    2    3    0    0    0
#> [545,]    2    3    0    0    0    0
#> [546,]    2    3    0    0    2    0
#> [547,]    2    3    1    0    0    0
#> [548,]    2    3    2    0    0    1
#> [549,]    3    0    0    0    0    0
#> [550,]    3    0    0    0    0    1
#> [551,]    3    0    0    0    0    2
#> [552,]    3    0    0    0    0    3
#> [553,]    3    0    0    0    2    0
#> [554,]    3    0    0    0    2    2
#> [555,]    3    0    0    0    3    0
#> [556,]    3    0    0    1    0    0
#> [557,]    3    0    0    2    0    0
#> [558,]    3    0    0    2    0    2
#> [559,]    3    0    0    2    2    0
#> [560,]    3    0    0    2    2    2
#> [561,]    3    0    0    3    0    0
#> [562,]    3    0    0    3    3    0
#> [563,]    3    0    1    0    0    0
#> [564,]    3    0    1    0    2    0
#> [565,]    3    0    1    0    3    0
#> [566,]    3    0    1    1    0    0
#> [567,]    3    0    2    0    0    0
#> [568,]    3    0    2    0    0    2
#> [569,]    3    0    2    2    0    0
#> [570,]    3    0    2    2    0    2
#> [571,]    3    0    2    2    2    0
#> [572,]    3    0    3    0    0    0
#> [573,]    3    0    3    0    0    3
#> [574,]    3    0    3    0    3    0
#> [575,]    3    0    3    0    3    3
#> [576,]    3    0    3    3    0    0
#> [577,]    3    0    3    3    0    2
#> [578,]    3    0    3    3    0    3
#> [579,]    3    0    3    3    3    0
#> [580,]    3    0    3    3    3    3
#> [581,]    3    1    0    0    0    0
#> [582,]    3    1    0    0    0    1
#> [583,]    3    1    0    0    0    2
#> [584,]    3    1    0    0    2    0
#> [585,]    3    1    0    0    2    1
#> [586,]    3    1    0    1    0    0
#> [587,]    3    1    0    1    1    1
#> [588,]    3    1    1    0    0    0
#> [589,]    3    1    1    0    1    0
#> [590,]    3    1    1    0    2    1
#> [591,]    3    1    1    1    0    0
#> [592,]    3    1    2    1    1    1
#> [593,]    3    2    0    0    0    0
#> [594,]    3    2    0    0    0    2
#> [595,]    3    2    0    0    2    0
#> [596,]    3    2    0    2    0    0
#> [597,]    3    2    0    2    1    2
#> [598,]    3    2    0    3    2    0
#> [599,]    3    2    1    0    0    0
#> [600,]    3    2    2    0    0    0
#> [601,]    3    2    2    0    0    2
#> [602,]    3    2    2    2    0    0
#> [603,]    3    2    2    2    0    1
#> [604,]    3    2    3    0    0    0
#> [605,]    3    3    0    0    0    0
#> [606,]    3    3    0    0    0    3
#> [607,]    3    3    0    0    3    0
#> [608,]    3    3    0    2    0    0
#> [609,]    3    3    0    2    3    0
#> [610,]    3    3    0    3    0    0
#> [611,]    3    3    1    0    0    0
#> [612,]    3    3    2    0    0    0
#> [613,]    3    3    2    0    0    2
#> [614,]    3    3    2    0    2    3
#> [615,]    3    3    2    2    0    0
#> [616,]    3    3    2    3    3    0
#> [617,]    3    3    3    0    0    0
#> [618,]    3    3    3    0    3    3
#> [619,]    3    3    3    2    0    2
#> [620,]    3    3    3    3    0    0
#> [621,]    3    3    3    3    0    1
#> [622,]    3    3    3    3    0    3
#> [623,]    3    3    3    3    3    3
#> 
#> $sample_size
#>   [1]  158  352  271  317   62   31  748   56  234    7  504   14   71   72  643
#>  [16]   57   36   98   42   10   70   13   25    1    1    2 1040   27  117   11
#>  [31]   59    2   12  285   11  100    2   10    1    5    2  694    7   66   60
#>  [46]   13    1    2   59    2   13    2   86    1   12   31 1135   40   39    2
#>  [61]   63   22    8   68    8   13    1    2    1  201   25   11   56   38    1
#>  [76]   21    4    1   81    9   12   11    5    4   22    6    4    1    3    1
#>  [91] 1559   18   85    4   33    6    3  163    3   51    1    7    1   55    6
#> [106]    2    9    1    4    7    1    4    1  423   10   45    5    5    4  121
#> [121]    2   48    6    1   16    4    1    1  473    1   17   21    9    1   19
#> [136]    2    1   20    1    7    6   22    3    5    1    1    2    1   57    2
#> [151]    2   15    3   10    3   13  920   18    9   36    9    1   24    5    6
#> [166]    3   83    7    1   22    7    2   44    2    7    1    1    5    2    2
#> [181]    2    1  195    8    2   29    8    1    5    1   52    6    1   41   24
#> [196]    2    5    6    2    1    1   73    3    3    2    1    4    2    8    1
#> [211]    1   19    3    2    1    1 1769   17   56    1   28    2    4  103    7
#> [226]   29    1    1   39    2    1    4    1    1    3    1    1  168    3   20
#> [241]    1    4    1   48   15    3    7    1    1   65    2    2    2    1   14
#> [256]    2    1   12    3    3    1    4    1    8    1    2    4    2    1    1
#> [271]  391    3   35    1    1    2   49    2   22    4    1   10    1    2    2
#> [286]   89    3   15    1    3   38   36    1    1    6    2    1    1    2    1
#> [301]  623    1   14    9    5    1    1   12    2    1    9    3    5    6    1
#> [316]    1   27    2    2    7    1    5   26    1    8    1    2    8    7    1
#> [331]    1    1    1    1   51    3    2    1    4    1    1   12    2    2    3
#> [346]    2    7    1    1   56    2    2    1    4    8    1    5    1    1    1
#> [361]    1   25   11    2    4   11  401    5    2    6    4    1    1    6    3
#> [376]    1    2    1    4    2   23    1    1    1    5    3    1    1    9    1
#> [391]    1    1    1    1    3    1  128    1    5    2   12    3    4    2    2
#> [406]    2   40    1    5    4    1    1   10    1    5    4    1    1    1    3
#> [421]   32    1    1    2    1    5    1    1    2    1    1    5 1001    3   22
#> [436]    3    3    1   26    1    4    1    2   18    1    2    1   60    1   11
#> [451]   16    4    1    1    1   19    1    2    2    1    1    1    6    1    1
#> [466]    1   82    2   12    1    1   11    4    3    1    1   23    1    7    6
#> [481]    1    1    1    1    1   47    1    3    1    1    1    1    1    1    1
#> [496]    1    1   10    1    4    1    1    1    1    6  328    1   16    1    5
#> [511]    1    1   18    1    3    1    5    2    1   41    1    6    3   11    1
#> [526]    5    1    9    1    1    2    2   73    1    4    7    1   39    1    5
#> [541]    1    6    8    1    3    1    1    1  380    2    5    3    6    2    3
#> [556]    3    7    1    1    1    7    3    8    1    1    1   19    2    7    1
#> [571]    1    9    1    1    1    2    1    1    1    1    9    1    1    2    1
#> [586]    1    1    4    1    1    1    1   36    1    2    1    1    1    1    2
#> [601]    1    1    1    1   29    2    1    2    1    2    1    2    1    1    2
#> [616]    1   10    1    1    4    1    2    3
#>